% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{adjustRT}
\alias{adjustRT}
\title{Peak alignment}
\usage{
adjustRT(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[xcms:adjustRtime]{xcms::adjustRtime}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
XCMSnExp object
}
\description{
\code{adjustRT} is a wrapper of the \code{xcms::adjustRtime} from \code{xcms} package. It
performs the \strong{alignment} of chromatographic peaks of samples. There are two
methods.
A) peakGroups: Correction of RT based on alignment of features (grouped
peaks) in most of samples. See \link[xcms:adjustRtime-peakGroups]{xcms::adjustRtime-peakGroups} for info.
B) obiwarp: Alignment based on the complete mz-rt data. This method does not
require any identified peaks or defined features. See
\link[xcms:adjustRtime-obiwarp]{xcms::adjustRtime-obiwarp} for info.
}
\seealso{
\code{\link[xcms]{retcor}} for the \emph{old} retention time correction
    methods.
    \code{\link[xcms]{plotAdjustedRtime}} for visualization of alignment results.

Other retention time correction methods: \code{\link[xcms]{adjustRtime-obiwarp}},
  \code{\link[xcms]{adjustRtime-peakGroups}}
}
\author{
Johannes Rainer
}
