% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retention_Time_Correction.R
\name{align_Rtime}
\alias{align_Rtime}
\title{Retention Time Correction}
\usage{
align_Rtime(peakdet, params)
}
\arguments{
\item{params}{A converted parameters template from IPO parameters.}
}
\description{
Retention time correction is performed  using *'obiwarp'* method.
Its optimum parameters are obtaine from `IPO` Package.
}
