% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB_metabolites.R
\name{assignation_neg_HMDB}
\alias{assignation_neg_HMDB}
\title{HMDB metabolites in negative ionization}
\usage{
assignation_neg_HMDB(feature_table)
}
\arguments{
\item{feature_table}{A data frame containing feature in rows and samples in
columns. First and second columns need to be the feature ID and mass.}
}
\value{
The same feature table with two extra column with matched HMDB codes
and metabolites. More than one assignation is allowed.
}
\description{
\code{assignation_neg_HMDB} is a function to identify the features in negative
ionization. It assumes deprotonation of molecules (M - H) from the feature
table and compares with mass of metabolites contained in the The Human
Metabolome Database \link{HMDB_db}.
}
\examples{
\dontrun{
FT <- assignation_neg_HMDB(feature_table)
head(FT)
}
}
