% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB_metabolites.R
\name{assignation_pos_HMDB}
\alias{assignation_pos_HMDB}
\title{HMDB metabolites in positive ionization}
\usage{
assignation_pos_HMDB(feature_table)
}
\arguments{
\item{feature_table}{A dataframe containing feature in rows and samples in
columns. First and second columns need to be the feature ID and mass.}
}
\value{
The same feature table with an extra column with matched metabolites
}
\description{
The function compares (M + H) mass of metabolites contained in the The Human
Metabolome Database (HMDB) with features from the positive ionized feature
table.
}
\examples{
\dontrun{
FT <- assignation_pos_HMDB(feature_table)
head(FT)
}
}
