% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{bp_kfold_VIP_analysis}
\alias{bp_kfold_VIP_analysis}
\title{K-fold bootstrap and permutation over PLS-VIP}
\usage{
bp_kfold_VIP_analysis(
  dataset,
  y_column,
  k = 4,
  ncomp = NULL,
  nbootstrap = 300,
  multilevel = NULL
)
}
\arguments{
\item{dataset}{An lcms_dataset object}

\item{y_column}{A list with the clases of the samples}

\item{k}{Number of folds, recomended between 4 to 10}

\item{ncomp}{number of components for the bootstrap models, if NULL it will
be automatically calculated}

\item{nbootstrap}{number of bootstrap dataset}
}
\value{
A list with the following elements:
\itemize{
\item \code{important_vips}: A list with the important vips selected
\item \code{relevant_vips}: List of vips with some relevance
\item \code{wilcoxon_vips}: List of vips that pass a wilcoxon test
\item \code{vip_means}: Means of the vips scores
\item \code{vip_score_plot}: plot of the vips scores
\item \code{kfold_resuls}: results of the k \link{bp_VIP_analysis}
\item \code{kfold_index}: list of index of partitions of the folds
}
}
\description{
Bootstrap and permutation over PLS-VIP on Alpslcms can be performed on both
\link{lcms_dataset_1D} full spectra as well as \link{lcms_dataset_peak_table} peak tables.
}
\details{
Use of the bootstrap and permutation methods for a more robust
variable importance in the projection metric for partial least
squares regression, in a k-fold cross validation
}
