% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{bp_VIP_analysis}
\alias{bp_VIP_analysis}
\title{Bootstrap and permutation over PLS-VIP}
\usage{
bp_VIP_analysis(
  dataset,
  train_index,
  y_column,
  ncomp,
  nbootstrap = 300,
  multilevel = NULL
)
}
\arguments{
\item{dataset}{An lcms_dataset object}

\item{train_index}{set of index used to generate the bootstrap datasets}

\item{y_column}{A list with the clases of the samples}

\item{ncomp}{number of components used in the plsda models}

\item{nbootstrap}{number of bootstrap dataset}
}
\value{
A list with the following elements:
\itemize{
\item \code{important_vips}: A list with the important vips selected
\item \code{relevant_vips}: List of vips with some relevance
\item \code{pls_vip}: Pls-VIPs of every bootstrap
\item \code{pls_vip_perm}: Pls-VIPs of every bootstrap with permuted variables
\item \code{pls_vip_means}: Pls-VIPs normaliced differences means
\item \code{pls_vip_score_diff}: Differences of \code{pls_vip} and \code{pls_vip_perm}
\item \code{pls_models}: pls models of the diferent bootstraps
\item \code{pls_perm_models}: pls permuted models of the diferent bootstraps
\item \code{classif_rate}: classification rate of the bootstrap models
\item \code{general_model}: pls model trained with all train data
\item \code{general_CR}: classification rate of the \code{general_model}
\item \code{vips_model}: pls model trained with vips selection over all train data
\item \code{vips_CR}: classification rate of the \code{vips_model}
\item \code{error}: error spected in a t distribution
\item \code{lower_bound}: lower bound of the confidence interval
\item \code{upper_bound}: upper bound of the confidence interval
}
}
\description{
Bootstrap and permutation over PLS-VIP performed on lcms_dataset peak tables.
}
\details{
Use of the bootstrap and permutation methods for a more robust
variable importance in the projection metric for partial least
squares regression
}
