% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPO_to_XCMS.R
\name{convert_IPO_to_XCMS}
\alias{convert_IPO_to_XCMS}
\title{Converts IPO parameters to XCMS format}
\usage{
convert_IPO_to_XCMS(params)
}
\arguments{
\item{params}{An object with the IPO parameters, generated with the \code{write_opt_params} function}
}
\value{
Parameters in XCMS format
}
\description{
The function converts from \code{IPO} to \code{XCMS} variable formats: The variable names in IPO and XCMS
presented some mismatches, thus we had to create a function called \emph{IPO_to_XCMS}
to achieve compatibility between packages.
}
\examples{
\dontrun{
write_opt_params(resultPeakpicking, resultRetcorGroup, my_directory)
}
}
