% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{do.findmain}
\alias{do.findmain}
\title{Cluster annotation}
\usage{
do.findmain(...)
}
\description{
Cluster annotation using the \link[InterpretMSSpectrum:findMain]{InterpretMSSpectrum::findMain}. It creates a
pseudo-spectrum with the features grouped in each cluster within a given mz
error. The algorithm proposes a representative ion of each
cluster/pseudo-spectrum for data reduction.
}
\section{}{

}

