% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{do_findmain}
\alias{do_findmain}
\title{do_findmain}
\usage{
do_findmain(
  ramclustObj = NULL,
  cmpd = NULL,
  mode = "positive",
  mzabs.error = 0.005,
  ppm.error = 10,
  ads = NULL,
  nls = NULL,
  scoring = "auto",
  plot.findmain = TRUE,
  writeMat = TRUE,
  writeMS = TRUE,
  use.z = TRUE
)
}
\description{
Cluster annotation function: inference of 'M' - molecular weight of the
compound giving rise to each spectrum - using the
InterpretMSSpectrum::findMain function
}
