% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{feature_reduction}
\alias{feature_reduction}
\title{Feature reduction}
\usage{
feature_reduction(mdata,
representative_ions,
RC)
}
\arguments{
\item{mdata}{data matrix with samples in rows and features in columns}

\item{representative_ions}{data frame from the list given by \link{labelling} function}

\item{RC}{hclust object from the \link{do_findmain} function}
}
\value{
A reduced data frame with important features based on the clustering
and features no clustered.
}
\description{
\code{feature_reduction} is a function to reduce the feature table using the info
from the \link{clustering}. A representative feature/ion is selected from each
cluster discarding the rest of the feature of the same cluster (normally, the
most intense one).
}
