% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{feature_values}
\alias{feature_values}
\title{Feature table}
\usage{
feature_values(
object,
method = c("medret", "maxint", "sum"),
value = "into",
intensity = "into",
filled = TRUE,
missing = NA,
msLevel = integer()
)
}
\value{
For \code{featureValues}: a \code{matrix} with
feature values, columns representing samples, rows features. The order
of the features matches the order found in the
\code{featureDefinitions(object)} \code{DataFrame}. The rownames of the
\code{matrix} are the same than those of the \code{featureDefinitions}
\code{DataFrame}. \code{NA} is reported for features without
corresponding chromatographic peak in the respective sample(s).
}
\description{
\code{feature_values} is a wrapper from \code{xcms}. The function creates an intensity
matrix with all the features.
}
\section{}{

}

\seealso{
\code{\link[xcms]{XCMSnExp}} for information on the data object.

\code{\link[xcms]{featureDefinitions}} to extract the \code{DataFrame} with the
feature definitions.

\code{\link[xcms]{featureChromatograms}} to extract ion chromatograms for each
feature.

\code{\link[xcms]{hasFeatures}} to evaluate whether the
\code{\link[xcms]{XCMSnExp}} provides feature definitions.

\code{\link[xcms]{groupval}} for the equivalent method on \code{xcmsSet} objects.
}
\author{
Johannes Rainer
}
