% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{featurevalues}
\alias{featurevalues}
\title{Feature table}
\usage{
featurevalues(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[xcms:XCMSnExp-peak-grouping-results]{xcms::featureValues}}
  \describe{
    \item{\code{object}}{A \code{\link[xcms]{XCMSnExp}} object providing the feature
definitions.}
  }}
}
\value{
A feature table
}
\description{
The function creates a intensity matrix with all the features.
}
\seealso{
\code{\link[xcms]{XCMSnExp}} for information on the data object.

\code{\link[xcms]{featureDefinitions}} to extract the \code{DataFrame} with the
feature definitions.

\code{\link[xcms]{featureChromatograms}} to extract ion chromatograms for each
feature.

\code{\link[xcms]{hasFeatures}} to evaluate whether the
\code{\link[xcms]{XCMSnExp}} provides feature definitions.

\code{\link[xcms]{groupval}} for the equivalent method on \code{xcmsSet} objects.
}
\author{
Johannes Rainer
}
