% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{filterMassAcquisition}
\alias{filterMassAcquisition}
\title{Filter a mass acquisition}
\usage{
filterMassAcquisition(dataset, window = "small")
}
\arguments{
\item{dataset}{A lcms_dataset that contains two different mass ranges in two different acquisitions.}

\item{window}{Character. It refers to which mass range will be kept}
}
\value{
A lcms_dataset that only contains one mass range
}
\description{
A function to filter by a mass acquisition when two different mass
acquisition have been acquired in the same experiment.
}
