% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_filtering.R
\name{filterPolarity}
\alias{filterPolarity}
\title{filter an experiment by its polarity}
\usage{
filterPolarity(object, polarity.)
}
\arguments{
\item{object}{An MSnExp object}

\item{polarity.}{The polarity to keep}
}
\description{
This function was merged in
the MSnbase package on 2018-12-10 as proposed at:
}
\details{
https://github.com/lgatto/MSnbase/issues/388

Once MSnbase is released (Bioconductor 3.9, maybe once R 3.6 is out),
remove this function.
}
