% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPolarity.R
\name{filterRTmin}
\alias{filterRTmin}
\title{Filter by retention time}
\usage{
filterRTmin(lcms_dataset, rt = c(4, 14))
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{rt}{Range of the retention time to keep in minutes}

\item{rt}{Retention time}

\item{plot_type}{The plot class, either boxplot or spectra}
}
\value{
A filtered dataset with the range of selected rt
}
\description{
The function converts seconds into minutes to cut and keep
a range of the retention time in minutes.
}
\seealso{
Other lcms_dataset functions: \code{\link{filterSampleType}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}},
  \code{\link{load_and_save_functions}}
}
\concept{lcms_dataset functions}
