% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_filtering.R
\name{filterSampleType}
\alias{filterSampleType}
\title{Filter by sample treatment}
\usage{
filterSampleType(lcms_dataset, especial_samples)
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{especial_samples}{A list with the especial samples names.
Use \code{NULL} if there is not any especial sample in the dataset.}
}
\value{
A filtered dataset with the remained samples
}
\description{
In a dataset, there are different types of samples for
checking purposes. For instance, quality control (QC) samples
may be pools of all samples, and blank samples may be composed
by the solvent used. Therefore, these samples should be removed
before alignment and processing from the main dataset. Use \code{NULL}
if there is not any especial sample in the dataset.
}
\details{
The function can distinguish and filter different samples types:
\itemize{
\item Regular samples
\item Blank samples
\item Quality Control samples
}
}
\examples{
\dontrun{
especial_samples <-list(QC = "DMSO-Ctrl", blank = "blank")
datasets_by_class_type <- filterSampleType(lcms_dataset, especial_samples)

lcms_dataset_no_blanks <-datasets_by_class_type$regular_samples
lcms_dataset_blanks <- datasets_by_class_type$blanks
lcms_dataset_QCs <- datasets_by_class_type$QCs
}#'
}
\seealso{
Other lcms_dataset functions: \code{\link{filterPolarity}},
  \code{\link{filterRTmin}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_filtering functions: \code{\link{filterPolarity}},
  \code{\link{filterRTmin}}
}
\concept{lcms_dataset functions}
\concept{lcms_filtering functions}
