% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{find_peaks_cwp}
\alias{find_peaks_cwp}
\title{Chromatographic peak detection (CentWave)}
\usage{
find_peaks_cwp(dataset, params)
}
\arguments{
\item{dataset}{A lcms_dataset.}

\item{params}{A converted parameters template from IPO parameters.}
}
\value{
A lcms_dataset with the detected peaks added.
}
\description{
The findChromPeaks_cwp function performs the chromatographic peak
detection on LC/GC-MS data. The standard method for peak detection
is \emph{'CentWave'}. Peak detection aims to detect important
features (peaks) on the chromatographic axis. This will be useful
for a posterior peak alignment on the chormatophic axis.
}
\details{
Note: signal processing generally  consists in three main steps: (1) peak
detection (\code{lcms_find_chromp_peaks_cwp} function), (2) peak alignment
(\code{obiAdjust_Rtime} function) and (3) peak correspondence (\code{IPO_group_peaks}
function). The optimized set of parameters for this signal preprocessing can
be obatained from \code{IPO} Package.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "rearrange_mait", package = "AlpsLCMS")
samples_mzxml <- list.files(file_path, recursive = TRUE, full.names = TRUE)
meta_path <- system.file("extdata", "metadata.xlsx", package = "AlpsLCMS")
opt_result_path <-  system.file("extdata", package = "AlpsLCMS")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

dataset <- suppressWarnings(lcms_read_samples(samples_mzxml, mode = "onDisk"))
metadata <- lcms_meta_read(meta_path)
dataset_meta <- lcms_meta_add(dataset, metadata, by = "sampleNames")

peakdet <- find_peaks_cwp(dataset_meta, params = preproc_params)
print(peakdet)
}
}
\seealso{
Other dataset functions: 
\code{\link{IPO_group_peaks}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_fill_chrom_peaks}()},
\code{\link{lcms_filter_mz}()},
\code{\link{lcms_filter_polarity}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()},
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_peakpicking_optimization}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_retention_time_alignment_plot}()},
\code{\link{lcms_tics}()},
\code{\link{obiAdjust_Rtime}()},
\code{\link{phData}()}
}
\concept{dataset functions}
\concept{peak detection functions}
