% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatographic_peak_detection.R
\name{findChromPeaks_cwp}
\alias{findChromPeaks_cwp}
\title{Chromatographic peak detection (CentWave)}
\usage{
findChromPeaks_cwp(lcms_dataset, params)
}
\arguments{
\item{lcms_dataset}{An [lcms_dataset_family] object}

\item{params}{A converted parameters template from IPO parameters.}
}
\description{
The findChromPeaks_cwp function performs the chromatographic peak
detection on LC/GC-MS data. The standard method for peak detection
is *'CentWave'*. We must initialize its parameters according to the
`IPO` Package optimization. Peak detection aims to detect important
features (peaks) on the chromatographic axis. This will be useful
for a posterior peak alignment on the chormatophic axis.
}
