% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getRamSt}
\alias{getRamSt}
\title{Get st parameter value for RamclustR}
\usage{
getRamSt(XObj)
}
\arguments{
\item{XObj}{A lcms_dataset}
}
\value{
recomended st parameter value for RamclustR
}
\description{
Get st parameter value for RamclustR
}
