% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_grouped_correspondence.R
\name{group_peaks}
\alias{group_peaks}
\title{Peak Correspondence}
\usage{
group_peaks(peakdet_align, params)
}
\arguments{
\item{peakdet_align}{A lcms_dataset with (1) detected (\code{findChromPeaks_cwp}
function) and (2) aligned (\code{align_Rtime} function) peaks}

\item{params}{A converted parameters template from IPO parameters.}
}
\value{
A lcms_dataset with (1) detected (\code{findChromPeaks_cwp} function),
(2) aligned (\code{align_Rtime} function) and (3) grouped  (\code{group_peaks}
function) peaks.
}
\description{
Peak correspondence is carried out by the \emph{'groupChromPeaks'} method,
with parameters obtained form \code{IPO}. Peak Correspondece consist in
grouping peaks on retention time axis with the purspose of associate
them to spectra on the mass/chage axis.
#' Note: signal processing generally  consists in three main steps:
(1) peak detection (\code{findChromPeaks_cwp} function),
(2) peak alignment (\code{align_Rtime} function) and
(3) peak correspondence (\code{group_peaks} function).
After this stage the peak table is finally obtained.
}
\examples{
\dontrun{
params <- convert_IPO_to_XCMS(IPO_params)
peakdet <- findChromPeaks_cwp(dataset, params = params)
peakdet_align <- align_Rtime(peakdet, params = params)
peak_table <- group_peaks(peakdet_align, params = params)
}
}
