% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB_metabolites.R
\docType{data}
\name{HMDB_db}
\alias{HMDB_db}
\title{The Human Metabolome DataBase template: Table with metabolites along with
their ionized masses}
\description{
The Human Metabolome DataBase template: Table with metabolites along with
their ionized masses
}
\references{
\url{hmdb.ca}
}
\keyword{data}
