% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_peak_annotation.R
\name{identify_metabolites}
\alias{identify_metabolites}
\title{Metabolite identifier}
\usage{
identify_metabolites(MAIT.object = NULL, peakTolerance = 0.005,
  database = NULL, polarity = "positive", printCSVfile = TRUE)
}
\arguments{
\item{MAIT.object}{
A \link{MAIT-class} object where significant features have already been found.
}

\item{peakTolerance}{
Maximum difference between the peak masses differences and the values shown in the database to be considered as a match. As default the value is 0.005 Da.
}

\item{database}{
User-defined input table. If it is set to NULL, the default \pkg{MAIT} database is selected to perform the metabolite identification.
}

\item{polarity}{
Character parameter that can be set to "positive" or "negative" depending on the polarity in which the samples were taken.
}

\item{printCSVfile}{
Set to TRUE if an output table has to be produced. The table should be found in (working directory)/Tables/SearchTable.csv. 
}
}
\value{
An output table is stored in the folder (working
directory)/Tables/SearchTable.csv if printCSVfile is set to TRUE. More info
at metaboliteTable
}
\description{
Takes a MAIT object and performs the metabolite search for the significant
features
}
\examples{
\dontrun{
peak_table_MAIT <- to_MAIT(dataDir = dataDir,
        project = project,
        preproc_params = preproc_params,
        peakTable = peakTable)

peak_table_ANN <- peak_annotation(MAIT.object = peak_table_MAIT)
peak_table_sig_ANN <- spectralSigFeatures(MAIT.object = peak_table_neg_ANN,
pvalue=0.05,
p.adj="none",
scale=FALSE)
sign_table <- sig_peaks_table(peak_table_sig_ANN)
Identif <- identifyMetabolites(MAIT.object = peak_table_sig_ANN,
                              peakTolerance = 0.005)
}
}
