% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_converter.R
\name{install_RawConverter}
\alias{install_RawConverter}
\title{Download and Install RawConverter}
\usage{
install_RawConverter(rawconverter, force = FALSE)
}
\arguments{
\item{rawconverter}{Working directory to install RawConverter.}

\item{force}{logical. Set to \code{TRUE} to install even if RawConverter already exists.}
}
\value{
the working directory
}
\description{
On non-windows systems, this function requires wine and winetricks to be installed
}
