% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_read_samples.R
\name{install_RawConverter_Linux}
\alias{install_RawConverter_Linux}
\title{Download and Install RawConverter}
\usage{
install_RawConverter_Linux(workdir, force = FALSE)
}
\arguments{
\item{workdir}{Working directory to install RawConverter}

\item{force}{logical. Set to `TRUE` to install even if RawConverter already exists}
}
\value{
the working directory
}
\description{
This function requires wine and winetricks to be installed
}
