% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{is.negative}
\alias{is.negative}
\title{is.negative}
\usage{
is.negative(dataset)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{
Function to confirm if the dataset was acquired in negative mode.
}
