% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{is.positive}
\alias{is.positive}
\title{is.positive}
\usage{
is.positive(dataset)
}
\arguments{
\item{dataset}{A lcms_dataset.}
}
\description{
Function to confirm if the dataset was acquired in positive mode.
}
