% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_metabolites.R
\name{known_metabolites}
\alias{known_metabolites}
\title{Known metabolites in the dataset}
\usage{
known_metabolites(feature_table_complete, metabolites)
}
\arguments{
\item{feature_table_complete}{A dataframe containing feature in rows and
samples in columns. First, second and third columns need to be the feature
ID, mass and retention time (min). column needs}

\item{metabolites}{A dataframe with metabolites containing a "mz" (mass) and
"rt" (retention time in min) columns.}
}
\value{
The same feature table with an extra column with matched metabolites
}
\description{
The function compares retention time and mass of metabolites contained in a
given template with features in the feature table.
}
\examples{
\dontrun{
FT <- known_metabolites(feature_table_complete, metabolites)
head(FT)
}

}
