% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{labelling}
\alias{labelling}
\title{Labelled fragments}
\usage{
labelling(RC)
}
\arguments{
\item{RC}{hclust object from the \link{do_findmain} function.}
}
\value{
A list containing data frames with (1) all labelled adducts, (2)
representative ion in each cluster/pseudospectrum, and (3) the clustering
data.
$labelled_adducts: data frame containing the mz, relative intensity of the
fragment in the corresponding cluster, charge, selected adduct and ppm
error.
$labelled_ion:
}
\description{
\code{labelling} is a function that applied to the \code{hclust} object from
\link{clustering} and \link{do_findmain}, creates a list with extracted info from each
cluster/pseudo-spectrum.
}
