% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{lcms_clustering}
\alias{lcms_clustering}
\title{clustering of peaks}
\usage{
lcms_clustering(
  ms,
  featdelim,
  st,
  sr,
  ExpDes,
  normalize,
  sampNameCol,
  fftempdir
)
}
\description{
Main clustering function for grouping features based on their analytical
behavior.
}
