% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_ipo_to_xcms.R
\name{lcms_convert_ipo_to_xcms}
\alias{lcms_convert_ipo_to_xcms}
\title{Converts IPO parameters to XCMS format}
\usage{
lcms_convert_ipo_to_xcms(params)
}
\arguments{
\item{params}{A list with the IPO parameters, generated with the \code{lcms_write_opt_params} function}
}
\value{
Parameters in XCMS format
}
\description{
The function converts from \code{IPO} to \code{XCMS} variable formats: The variable names in IPO and XCMS
presented some mismatches, thus we had to create a function called \emph{lcms_convert_ipo_to_xcms}
to achieve compatibility between packages when these parameters are stored in a list.
}
\examples{
opt_result_path <-  system.file("extdata", "params.rds", package = "NIHSlcms")
lcms_preproc_params <- lcms_convert_ipo_to_xcms(params)
print(lcms_preproc_params)
}
