% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{lcms_data_analysis}
\alias{lcms_data_analysis}
\title{Data analysis}
\usage{
lcms_data_analysis(
  dataset,
  y_column,
  identity_column,
  balance_in_train = NULL,
  external_val,
  internal_val,
  data_analysis_method
)
}
\arguments{
\item{dataset}{An \link{lcms_dataset_family} object}

\item{y_column}{A string with the name of the y column (present in the
metadata of the dataset)}

\item{identity_column}{\code{NULL} or a string with the name of the identity column (present in the
metadata of the dataset).}

\item{external_val, internal_val}{A list with two elements: \code{iterations} and \code{test_size}.
See \link{random_subsampling} for further details}

\item{data_analysis_method}{An \link{lcms_data_analysis_method} object}
}
\value{
A list with the following elements:
\itemize{
\item \code{train_test_partitions}: A list with the indices used in train and test on each of the cross-validation iterations
\item \code{inner_cv_results}: The output returned by \code{train_evaluate_model} on each inner cross-validation
\item \code{inner_cv_results_digested}: The output returned by \code{choose_best_inner}.
\item \code{outer_cv_results}: The output returned by \code{train_evaluate_model} on each outer cross-validation
\item \code{outer_cv_results_digested}: The output returned by \code{train_evaluate_model_digest_outer}.
}
}
\description{
Data analysis on AlpsLCMS can be performed on both \link{lcms_dataset_1D} full spectra
as well as \link{lcms_dataset_peak_table} peak tables.
}
\details{
The workflow consists of a double cross validation strategy using random
subsampling for splitting into train and test sets. The classification model
and the metric to choose the best model can be customized (see
\code{\link[=new_lcms_data_analysis_method]{new_lcms_data_analysis_method()}}), but for now only a PLSDA classification
model with a best area under ROC curve metric is implemented (see
the examples here and \link{plsda_auroc_vip_method})
}
