% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_family.R
\name{lcms_dataset_family}
\alias{lcms_dataset_family}
\title{lcms_dataset like objects (S3 classes)}
\description{
The NIHSlcms package defines and uses several objects to manage LCMS Data.
}
\details{
These objects share some structure and functions, so it makes sense to have
an abstract class to ensure that the shared structures are compatible
}
\seealso{
\link[=load_and_save_functions]{Functions to save and load these objects}
}
\concept{NIHSnmr dataset objects}
