% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{lcms_dataset_load}
\alias{lcms_dataset_load}
\title{Function to load a lcms_dataset object}
\usage{
lcms_dataset_load(file_name)
}
\arguments{
\item{file_name}{The file name to load}
}
\value{
An object from the \link{lcms_dataset_family}
}
\description{
Function to load a lcms_dataset object
}
\examples{
file_name <- system.file("extdata", "dataset.rds", package = "NIHSlcms")
dataset <- lcms_dataset_load(file_name)
print(dataset)
}
\seealso{
Other dataset functions: \code{\link{lcms_dataset_save}},
  \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_export}}, \code{\link{phData}}

Other import/export functions: \code{\link{lcms_dataset_save}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_read_samples}}, \code{\link{phData}}
}
\concept{dataset functions}
\concept{dataset_peak_table functions}
\concept{import/export functions}
