% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_diagnose.R
\name{lcms_diagnose}
\alias{lcms_diagnose}
\alias{lcms_diagnose<-}
\title{Set/Get diagnostic information}
\usage{
lcms_diagnose(lcms_dataset)

lcms_diagnose(lcms_dataset) <- value
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{value}{The diagnostic we want to set}
}
\value{
Diagnostic information, usually a list with data or plots
}
\description{
Set or get diagnostic information
}
