% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_fillChromPeaks.R
\name{lcms_fillChromPeaks}
\alias{lcms_fillChromPeaks}
\title{Filling missing values in a peak tabe}
\usage{
lcms_fillChromPeaks(peak_table)
}
\arguments{
\item{peak_table}{A table of peaks with (possibly) missing values.}
}
\value{
A peak table where the missing peaks have been filled
}
\description{
In the imputation stage, we integrate the areas of the missing peaks of the peak table
that were not detected in the previous steps of the signal preprocessing workflow.
}
\examples{
file_name <-  system.file("extdata", "peak_table.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name)
peak_table_imputed <- lcms_fillChromPeaks(peak_table)

print(peak_table_imputed)
}
