% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_filter_mz}
\alias{lcms_filter_mz}
\title{Filter by mass/charge}
\usage{
lcms_filter_mz(dataset, mz)
}
\arguments{
\item{dataset}{A lcms_dataset.}

\item{mz}{The range of masses to filter.}
}
\value{
A filtered lcms_dataset with the selected m/z range.
}
\description{
The function filters the dataset by m/z.
}
\examples{
\dontrun{
mz_range = c(300, 400)
dataset <- lcms_dataset_load(system.file("extdata",
                                       "dataset_pos.rds",
                                       package = "AlpsLCMS"))
dataset_pos_mz <-lcms_filter_mz(dataset, mz = mz_range)

print(dataset_pos_mz)
}
}
\seealso{
Other dataset functions: 
\code{\link{lcms_align_rtime}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_fill_chrom_peaks}()},
\code{\link{lcms_filter_polarity}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()},
\code{\link{lcms_find_chrom_peaks_cwp}()},
\code{\link{lcms_group_peaks}()},
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_peakpicking_optimization}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_retention_time_alignment_plot}()},
\code{\link{lcms_tics}()},
\code{\link{phData}()}

Other filtering functions: 
\code{\link{lcms_filter_polarity}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()}
}
\concept{dataset functions}
\concept{filtering functions}
