% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_filtering.R
\name{lcms_filter_polarity}
\alias{lcms_filter_polarity}
\title{Filter an experiment by its polarity}
\usage{
lcms_filter_polarity(object, polarity.)
}
\arguments{
\item{object}{An MSnExp object}

\item{polarity.}{The polarity to keep}
}
\value{
A filtered \link{lcms_dataset_family} object with the selected polarity
}
\description{
This function was merged in
the MSnbase package on 2018-12-10 as proposed at:
}
\details{
https://github.com/lgatto/MSnbase/issues/388

Once MSnbase is released (Bioconductor 3.9, maybe once R 3.6 is out),
remove this function.
}
\examples{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata","dataset_metadata.rds",package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_filtering functions: \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}}
}
\concept{lcms_dataset functions}
\concept{lcms_filtering functions}
