% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_filter_rt_min}
\alias{lcms_filter_rt_min}
\title{Filter by retention time}
\usage{
lcms_filter_rt_min(dataset, rt = c(4, 14))
}
\arguments{
\item{dataset}{An \link{lcms_dataset_family} object filtered by retention time}

\item{rt}{Range of the retention time to keep in minutes}
}
\value{
A filtered dataset with the range of selected rt
}
\description{
The function converts seconds into minutes to cut and keep
a range of the retention time in minutes.
}
\examples{

rtime_range = c(5,10)
dataset <- lcms_dataset_load(system.file("extdata","dataset_pos.rds",package = "NIHSlcms"))
dataset_pos_rt <-lcms_filter_rt_min(dataset, rt = rtime_range)

print(dataset_pos_rt)

}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_filtering functions: \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_sample_type}}
}
\concept{lcms_dataset functions}
\concept{lcms_filtering functions}
