% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_filter_sample_type}
\alias{lcms_filter_sample_type}
\title{Filter by sample treatment}
\usage{
lcms_filter_sample_type(dataset, especial_samples)
}
\arguments{
\item{dataset}{A lcms_dataset}

\item{especial_samples}{A list with the especial samples names.
Use \code{NULL} if there is not any especial sample in the dataset.}
}
\value{
A list containing three lcms_datasets with different type: regular,
quality control and blank samples.
}
\description{
In a dataset, there are different types of samples for
checking purposes. For instance, quality control (QC) samples
may be pools of all samples, and blank samples may be composed
by the solvent used. Therefore, these samples should be removed
before alignment and processing from the main dataset. Use \code{NULL}
if there is not any especial sample in the dataset.
}
\details{
The function can distinguish and filter different samples types:
\itemize{
\item Regular samples.
\item Blank samples.
\item Quality Control samples.
}
}
\examples{
dataset <- lcms_dataset_load(system.file("extdata","dataset_pos_rt.rds",package = "NIHSlcms"))
especial_samples <- list(QC = NULL, blank = NULL)
datasets_by_class_type <- lcms_filter_sample_type(dataset, especial_samples)
dataset_pos_rt_rs <- datasets_by_class_type$regular_samples
dataset_pos_rt_qcs <- datasets_by_class_type$QCs
print(dataset_pos_rt_rs)

print(dataset_pos_rt_qcs)
}
\seealso{
Other dataset functions: \code{\link{lcms_align_rtime}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_fill_chrom_peaks}},
  \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_find_chrom_peaks_cwp}},
  \code{\link{lcms_group_peaks}},
  \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_rearrange_datafiles_by_class}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lcms_tics}},
  \code{\link{lmcs_plot_chrom_peak_image}},
  \code{\link{phData}}

Other filtering functions: \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}}
}
\concept{dataset functions}
\concept{filtering functions}
