% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_filtering.R
\name{lcms_filterPolarity}
\alias{lcms_filterPolarity}
\title{Filter an experiment by its polarity}
\usage{
lcms_filterPolarity(object, polarity.)
}
\arguments{
\item{object}{An MSnExp object}

\item{polarity.}{The polarity to keep}
}
\description{
This function was merged in
the MSnbase package on 2018-12-10 as proposed at:
}
\details{
https://github.com/lgatto/MSnbase/issues/388

Once MSnbase is released (Bioconductor 3.9, maybe once R 3.6 is out),
remove this function.
}
\examples{

lcms_dataset_2_polarities <- lcms_dataset_load(system.file("extdata","lcms_dataset_metadata.rds",package = "NIHSlcms"))
lcms_dataset_pos <- lcms_filterPolarity(lcms_dataset_2_polarities, polarity. = 1)

}
\seealso{
Other lcms_dataset functions: \code{\link{filterRTmin}},
  \code{\link{filterSampleType}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_filtering functions: \code{\link{filterRTmin}},
  \code{\link{filterSampleType}}
}
\concept{lcms_dataset functions}
\concept{lcms_filtering functions}
