% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_filtering.R
\name{lcms_filterRTmin}
\alias{lcms_filterRTmin}
\title{Filter by retention time}
\usage{
lcms_filterRTmin(lcms_dataset, rt = c(4, 14))
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{rt}{Range of the retention time to keep in minutes}
}
\value{
A filtered dataset with the range of selected rt
}
\description{
The function converts seconds into minutes to cut and keep
a range of the retention time in minutes.
}
\examples{

rtime_range = c(4,14)
lcms_dataset <- lcms_dataset_load(system.file("extdata","lcms_dataset_metadata.rds",package = "NIHSlcms"))
lcms_dataset_rt <-lcms_filterRTmin(lcms_dataset, rt = rtime_range)

}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_filterPolarity}},
  \code{\link{lcms_filterSampleType}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_filtering functions: \code{\link{lcms_filterPolarity}},
  \code{\link{lcms_filterSampleType}}
}
\concept{lcms_dataset functions}
\concept{lcms_filtering functions}
