% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_chromatographic_peak_detection.R
\name{lcms_findChromPeaks_cwp}
\alias{lcms_findChromPeaks_cwp}
\title{Chromatographic peak detection (CentWave)}
\usage{
lcms_findChromPeaks_cwp(lcms_dataset, params)
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{params}{A converted parameters template from IPO parameters.}
}
\value{
A lcms_dataset with detected peaks
}
\description{
The findChromPeaks_cwp function performs the chromatographic peak
detection on LC/GC-MS data. The standard method for peak detection
is \emph{'CentWave'}. We must initialize its parameters according to the
\code{IPO} Package optimization. Peak detection aims to detect important
features (peaks) on the chromatographic axis. This will be useful
for a posterior peak alignment on the chormatophic axis.
Note: signal processing generally  consists in three main steps:
(1) peak detection (\code{lcms_findChromPeaks_cwp} function),
(2) peak alignment (\code{lcms_align_Rtime} function) and
(3) peak correspondence (\code{lcms_group_peaks} function).
}
\examples{
file_path <- system.file("extdata", "rearrange_mait", package = "NIHSlcms")
samples_mzxml <- list.files(file_path, recursive = TRUE, full.names = TRUE)
meta_path <- system.file("extdata", "metadata.xlsx", package = "NIHSlcms")
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
lcms_preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

lcms_dataset <- suppressWarnings(lcms_read_samples(samples_mzxml, mode = "onDisk"))
metadata <- lcms_meta_read(meta_path)
lcms_dataset_meta <- lcms_meta_add(lcms_dataset, metadata, by = "sampleNames")

peakdet <- lcms_findChromPeaks_cwp(lcms_dataset_meta, params = lcms_preproc_params)
print(peakdet)

}
