% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_peak_grouped_correspondence.R
\name{lcms_group_peaks}
\alias{lcms_group_peaks}
\title{Peak Correspondence}
\usage{
lcms_group_peaks(peakdet_align, params)
}
\arguments{
\item{peakdet_align}{A lcms_dataset with (1) detected (\code{findChromPeaks_cwp}
function) and (2) aligned (\code{align_Rtime} function) peaks}

\item{params}{A converted parameters template from IPO parameters.}
}
\value{
A lcms_dataset with (1) detected (\code{findChromPeaks_cwp} function),
(2) aligned (\code{align_Rtime} function) and (3) grouped  (\code{group_peaks}
function) peaks.
}
\description{
Peak correspondence is carried out by the \emph{'lcms_groupChromPeaks'} method,
with parameters obtained form \code{IPO}. Peak Correspondece consist in
grouping peaks on retention time axis with the purspose of associate
them to spectra on the mass/chage axis.
#' Note: signal processing generally  consists in three main steps:
(1) peak detection (\code{lcmfindChromPeaks_cwp} function),
(2) peak alignment (\code{align_Rtime} function) and
(3) peak correspondence (\code{group_peaks} function).
After this stage the peak table is finally obtained.
}
\examples{
file_name <-  system.file("extdata", "peakdet_align.rds", package = "NIHSlcms")
peakdet_align <- base::readRDS(file_name)
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
lcms_preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

peak_table <- lcms_group_peaks(peakdet_align, params = lcms_preproc_params)
print(peak_table)

}
