% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_peak_annotation.R
\name{lcms_identify_metabolites}
\alias{lcms_identify_metabolites}
\title{Metabolite identifier}
\usage{
lcms_identify_metabolites(MAIT.object = NULL, peakTolerance = 0.005,
  database = NULL, polarity = "positive", printCSVfile = TRUE)
}
\arguments{
\item{MAIT.object}{
A \link{MAIT-class} object where significant features have already been found.
}

\item{peakTolerance}{
Maximum difference between the peak masses differences and the values shown in the database to be considered as a match. As default the value is 0.005 Da.
}

\item{database}{
User-defined input table. If it is set to NULL, the default \pkg{MAIT} database is selected to perform the metabolite identification.
}

\item{polarity}{
Character parameter that can be set to "positive" or "negative" depending on the polarity in which the samples were taken.
}

\item{printCSVfile}{
Set to TRUE if an output table has to be produced. The table should be found in (working directory)/Tables/SearchTable.csv. 
}
}
\value{
An output table is stored in the folder (working
directory)/Tables/SearchTable.csv if printCSVfile is set to TRUE. More info
at metaboliteTable
}
\description{
Takes a MAIT object and performs the metabolite search for the significant
features
}
\examples{
\dontrun{
file_name <- system.file("extdata", "peak_table_sig_ANN.rds", package = "NIHSlcms")
peak_table <- readRDS(file_name)
metabololite_table <- lcms_identify_metabolites(MAIT.object = peak_table,
                              peakTolerance = 0.005)
}
}
