% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_converter.R
\name{lcms_list_mzxml_samples}
\alias{lcms_list_mzxml_samples}
\title{RAW converter}
\usage{
lcms_list_mzxml_samples(
  sample_path,
  file_format = "mzXML",
  rawconverter_path = NULL
)
}
\arguments{
\item{sample_path}{Directory in which the samples are.}

\item{file_format}{Format of the LC-MS files (e.g. file_format = "raw").}

\item{rawconverter_path}{Directory in where the RawConverter application is located.}
}
\value{
A list of the LC-MS files in a readable format to create the \code{lcms_dataset}.
}
\description{
The function lists and converts samples in ".raw" format to ".mzXML" format.
It is the first step to create a \code{lcms_dataset}. It requires the previous
download and installation of the \emph{RawConverter} application.
}
\examples{
\dontrun{
sample_path <- system.file("extdata", package = "NIHSlcms")
samples_mzxml <- lcms_list_mzxml_samples(sample_path,
                                    file_format = "mzXML")
print(samples_mzxml)
}
}
