% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{lcms_meta_export}
\alias{lcms_meta_export}
\title{Export Metadata to an Excel file}
\usage{
lcms_meta_export(dataset, xlsx_file)
}
\arguments{
\item{dataset}{An \link{lcms_dataset_family} object}

\item{xlsx_file}{"The .xlsx excel file"}

\item{groups}{A character vector. Use \code{"external"} for the external metadata or
the default for a more generic solution}
}
\value{
The Excel file name
}
\description{
Export Metadata to an Excel file
}
\examples{
dataset_metadata <- lcms_dataset_load(system.file("extdata",
                                             "dataset_metadata.rds",
                                              package = "NIHSlcms"))
xlsx_file <- paste0(system.file("extdata",package = "NIHSlcms"),
                    "/", "exported_metadata.xlsx")

lcms_meta_export(dataset_metadata, xlsx_file)
print(dataset_metadata)

}
\seealso{
Other metadata functions: \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_read}}, \code{\link{phData}}

Other dataset functions: \code{\link{lcms_align_rtime}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_fill_chrom_peaks}},
  \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_find_chrom_peaks_cwp}},
  \code{\link{lcms_group_peaks}},
  \code{\link{lcms_meta_add}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_rearrange_datafiles_by_class}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lcms_tics}},
  \code{\link{lmcs_plot_chrom_peak_image}},
  \code{\link{phData}}

Other import/export functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_read_samples}},
  \code{\link{lcms_rearrange_datafiles_by_class}},
  \code{\link{phData}}
}
\concept{dataset functions}
\concept{import/export functions}
\concept{metadata functions}
