% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_load_save.R
\name{lcms_meta_read}
\alias{lcms_meta_read}
\title{Reads metadata from an Excel file}
\usage{
lcms_meta_read(xlsx_file)
}
\arguments{
\item{xlsx_file}{xlsx_file "The .xlsx excel file" with metadata}
}
\value{
A dataframe with the metadata
}
\description{
Reads metadata from an Excel file
}
\examples{

metadata <- lcms_meta_read(system.file("extdata",
                                             "metadata.xlsx",
                                              package = "NIHSlcms"))

}
\seealso{
Other metadata functions: \code{\link{lcms_meta_export}}

Other lcms_dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_filterPolarity}},
  \code{\link{lcms_filterRTmin}},
  \code{\link{lcms_filterSampleType}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other import/export functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_export}}
}
\concept{import/export functions}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
\concept{metadata functions}
