% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{lcms_peak_table_boxplots}
\alias{lcms_peak_table_boxplots}
\title{Boxplots for significant peak table features}
\usage{
lcms_peak_table_boxplots(MAIT.object = NULL, treatment_col)
}
\arguments{
\item{MAIT.object}{MAIT object where it is found an annotated peak table}

\item{treament_col}{Treatment for the samples.}
}
\value{
BoxPlots are stored in folders associated to their corresponding spectra. No explicit plot is produced by the device
}
\description{
It performs boxplots for any of significant features found in a peak table. All the
boxplots are stored in a directory /Boxplots/Boxplot_spectra_
}
\examples{
file_name_1 <-  system.file("extdata","peak_table_sig_ann.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name_1)
file_name_2 <-  system.file("extdata","dataset_pos_rt_rs.rds", package = "NIHSlcms")
dataset <-  base::readRDS(file_name_2)
treatment_col <- scales::hue_pal()(length(unique(dataset$treatment)))
lcms_peak_table_boxplots(peak_table,
                         treatment_col = treatment_col)
}
