% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{lcms_peak_table_pca}
\alias{lcms_peak_table_pca}
\title{Principal Component Analysis (PCA)}
\usage{
lcms_peak_table_pca(MAIT.object = NULL, treatment_col, Log = FALSE,
  center = TRUE, scale = TRUE)
}
\arguments{
\item{MAIT.object}{MAIT object where it is found an annotated peak table.}

\item{Log}{Set to TRUE if the data should be plotted using the logarithm of the intensity.}

\item{scale}{Set to TRUE if the data should be scaled.}

\item{treament_col}{Treatment for the samples.}

\item{Set}{to TRUE if the data should be centered around its mean. See scale.}
}
\value{
A MAIT.object. Additionaly: Three different PCA scoreplots are printed in three png files.
One using PC1 vs PC2, another with PC1 vs PC3 and the last one with PC2 vs PC3.
The files will be stored in the directory /PCA_Scoreplots.
#' @family metabolite identification functions
}
\description{
It performs PCA using on the annotated peak table obtained from a MAIT object.
}
\examples{
file_name_1 <-  system.file("extdata","peak_table_sig_ann.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name_1)
file_name_2 <-  system.file("extdata","dataset_pos_rt_rs.rds", package = "NIHSlcms")
dataset <-  base::readRDS(file_name_2)
treatment_col <- scales::hue_pal()(length(unique(dataset$treatment)))

lcms_peak_table_pca(peak_table,
                   treatment_col = treatment_col,
                   Log = FALSE,
                   center = TRUE, scale = FALSE)
}
\seealso{
Other dataset_peak_table functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_peak_table_boxplots}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_spectral_sig_features}},
  \code{\link{lcms_tics}}, \code{\link{lcms_to_mait}}

Other statistical functions: \code{\link{lcms_peak_table_boxplots}}

Other visualization functions: \code{\link{lcms_peak_table_boxplots}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_chrom}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_retcorgroup_optimization}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lmcs_plot_chrom_peak_image}}
}
\concept{dataset_peak_table functions}
\concept{statistical functions}
\concept{visualization functions}
