% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_plot_chrom}
\alias{lcms_plot_chrom}
\title{Base peak chromatogram}
\usage{
lcms_plot_chrom(chromatogram_object, treatment_col, rtlim = NULL)
}
\arguments{
\item{chromatogram_object}{A XChromatograms object}

\item{treatment_col}{Color code by groups.}

\item{rtlim}{retention time boundaries (e.g. c(4,8))}
}
\value{
A base peak chromatogram
}
\description{
Base peak chromatograms with retention time axis in minutes.
}
\examples{
file_name <- system.file("extdata",
                         "dataset_pos.rds",
                          package = "NIHSlcms")
dataset <- lcms_dataset_load(file_name)
treatment_col <- scales::hue_pal()(length(unique(dataset$treatment)))
names(treatment_col) <- unique(dataset$treatment)
base_peaks <- xcms::chromatogram(dataset, aggregationFun = "max")
lcms_plot_chrom(base_peaks, treatment_col, rtlim = c(4, 8))

}
\seealso{
Other chromatogram functions: \code{\link{lcms_plot_tics}},
  \code{\link{lmcs_plot_chrom_peak_image}}

Other visualization functions: \code{\link{lcms_peak_table_boxplots}},
  \code{\link{lcms_peak_table_pca}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_retcorgroup_optimization}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lmcs_plot_chrom_peak_image}}
}
\concept{chromatogram functions}
\concept{visualization functions}
