% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIC.R
\name{lcms_plot_tics}
\alias{lcms_plot_tics}
\title{Total Ion Count (TIC) plot}
\usage{
lcms_plot_tics(tics, treatment = treatment, rt = NULL,
  plot_type = "spec")
}
\arguments{
\item{tics}{A Total Ion Count object generated with `lcms_tics`}

\item{treatment}{Class groups of the samples}

\item{rt}{Retention time}

\item{plot_type}{The plot class, either boxplot or spectra}
}
\value{
Total Ion Count (TIC) for the polarity samples.
}
\description{
The function performs a plot with the Total Ion Count (TIC).
}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_meta_export}},
  \code{\link{lcms_tics}},
  \code{\link{load_and_save_functions}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_meta_export}},
  \code{\link{lcms_tics}},
  \code{\link{load_and_save_functions}}
}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
