% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_plotChrom.R
\name{lcms_plotChrom}
\alias{lcms_plotChrom}
\title{Base peak chromatogram}
\usage{
lcms_plotChrom(chromatogram, treatment_col, rtlim = NULL)
}
\arguments{
\item{chromatogram}{A XChromatograms object}

\item{treatment_col}{Color code by groups. They can be generated using
treatment_col <- scales::hue_pal()(length(unique(lcms_dataset$treatment)))
names(treatment_col) <- unique(lcms_dataset$treatment)}

\item{rtlim}{retention time boundaries (e.g. c(4,14))}
}
\value{
A base peak chromatogram
}
\description{
Base peak chromatograms with retention time axis in minutes.
}
\examples{

file_name <- system.file("extdata",
                         "lcms_dataset_pos.rds",
                          package = "NIHSlcms")
lcms_dataset <- lcms_dataset_load(file_name)
treatment_col <- scales::hue_pal()(length(unique(lcms_dataset$treatment)))
names(treatment_col) <- unique(lcms_dataset$treatment)
base_peaks <- xcms::chromatogram(lcms_dataset, aggregationFun = "max")
lcms_plotChrom(base_peaks, treatment_col, rtlim = c(4, 14))


}
