% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_read_samples.R
\name{lcms_raw_to_mzxml}
\alias{lcms_raw_to_mzxml}
\title{Convert RAW samples to mzxml samples}
\usage{
lcms_raw_to_mzxml(samples, options = c("--mzxml", "--select_mono_prec"),
  rawconverter = NULL)
}
\arguments{
\item{samples}{ThermoFisher raw files}

\item{rawconverter}{On Linux, the workdir for `install_RawConverter_Linux`. On
Windows, the path to RawConverter. See the RawConverter section.}
}
\description{
Convert RAW samples to mzxml samples
}
\section{RawConverter}{

Raw samples from ThermoFisher are digitally encrypted and cannot be read by
any software, allegedly to prevent sample forgery in pharmaceutical applications.

This function uses a tool named RawConverter available at:
http://fields.scripps.edu/rawconv/

RawConverter uses the ThermoFisher tool MSFileReader that takes care of the
decryption.

Installing that tool may not be straightforward on all OS, so we provide some
helpers.

On Linux systems:
  You will need wine and winetricks installed. Then call `install_RawConverter_Linux("/your/installation/directory")`
On Windows systems:
  You will need to install RawConverter yourself.

On OSX, the Linux solution may work, assuming you can install wine and winetricks.
}

