% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_read_samples.R
\name{lcms_read_samples}
\alias{lcms_read_samples}
\title{Read LCMS samples on mzXML format}
\usage{
lcms_read_samples(...)
}
\arguments{
\item{...}{Arguments passed on to \code{MSnbase::readMSData}
\describe{
  \item{files}{A \code{character} with file names to be read and parsed.}
  \item{pdata}{An object of class \linkS4class{AnnotatedDataFrame} or
\code{NULL} (default).}
  \item{msLevel.}{MS level spectra to be read. In \code{inMemory} mode,
use \code{1} for MS1 spectra or any larger numeric for MSn
spectra. Default is \code{2} for \code{InMemory} mode. \code{onDisk} mode
supports multiple levels and will, by default, read all the
data.}
  \item{verbose}{Verbosity flag. Default is to use
\code{\link[=isMSnbaseVerbose]{isMSnbaseVerbose()}}.}
  \item{centroided.}{A \code{logical}, indicating whether spectra are
centroided or not. Default is \code{NA} in which case the information
is extracted from the raw file (for mzML or mzXML files). In
\code{onDisk}, it can also be set for different MS levels by a
vector of logicals, where the first element is for MS1, the
second element is for MS2, ... See \linkS4class{OnDiskMSnExp} for
an example.}
  \item{smoothed.}{A \code{logical} indicating whether spectra already
smoothed or not. Default is \code{NA}.}
  \item{cache.}{Numeric indicating caching level. Default is 0 for
MS1 and 1 MS2 (or higher). Only relevant for \code{inMemory} mode.}
  \item{mode}{On of \code{"inMemory"} (default) or \code{"onDisk"}. The former
loads the raw data in memory, while the latter only generates
the object and the raw data is accessed on disk when
needed. See the \emph{benchmarking} vignette for memory and speed
implications.}
}}
}
\value{
An \linkS4class{MSnExp} object for \code{inMemory} mode and a
\linkS4class{OnDiskMSnExp} object for \code{onDisk} mode.
}
\description{
Read LCMS samples on mzXML format
}
\examples{
data(samples_mzxml)
lcms_dataset <- suppressWarnings(lcms_read_samples(samples_mzxml, mode = "onDisk"))
print(lcms_dataset)


}
