% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_rearrange_datafiles_by_class}
\alias{lcms_rearrange_datafiles_by_class}
\title{Rearrange datafiles by class}
\usage{
lcms_rearrange_datafiles_by_class(dataset, dataDir)
}
\arguments{
\item{dataset}{A lcms_dataset.}

\item{dataDir}{A directory in where LC-MS files are
going to be saved.}
}
\value{
LC-MS datafiles sorted by class treatment.
}
\description{
Although Data Preprocessing can be performed using any
filepath with the \code{XCMS} Package, it is convenient to
rearrange data files by sample class (that is, all the
samples belonging to the same sample class are included
in the same folder). This is done because, after the
preprocessing stage, \code{MAIT} Package takes care
of Data Annotation and Metabolite Identification. \code{MAIT}
needs a specific directory structure for data managing.
Otherwise, it can't work properly.
}
\examples{
\dontrun{
file_name <- system.file("extdata", "dataset_pos_rt_rs.rds", package = "NIHSlcms")
dataset <- lcms_dataset_load(file_name)
path <- system.file("extdata","rearrange_mait", package = "NIHSlcms")
lcms_rearrange_datafiles_by_class(dataset = dataset,
                            dataDir = path)
fileList = list.files(path = paste(path, list.files(path = path),
                           sep = "/"), full.names = TRUE)
print(fileList)
}
}
\seealso{
Other dataset functions: \code{\link{lcms_align_rtime}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_fill_chrom_peaks}},
  \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_find_chrom_peaks_cwp}},
  \code{\link{lcms_group_peaks}},
  \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_chrom_peak_image}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lcms_tics}}, \code{\link{phData}}

Other import/export functions: \code{\link{lcms_convert_ipo_to_xcms}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_read_ipo_to_xcms}},
  \code{\link{lcms_read_samples}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_spectral_sig_features}},
  \code{\link{lcms_to_mait}},
  \code{\link{lcms_write_opt_params}},
  \code{\link{lcms_write_parameter_table}},
  \code{\link{phData}}
}
\concept{dataset functions}
\concept{import/export functions}
