% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_parameter_optimization.R
\name{lcms_retcorgroup_optimization}
\alias{lcms_retcorgroup_optimization}
\title{Optimization of retention time correction and grouping parameters.}
\usage{
lcms_retcorgroup_optimization(optimizedXcmsSetObject,
  retcorGroupParameters, nSlaves = 1, opt_path, plots = TRUE)
}
\arguments{
\item{optimizedXcmsSetObject}{XCMS object conatining the \code{best_settings} parameters.
This object may be created after running \code{lcms_peakpicking_optimization} and extract
best_settings$xset (e.g. optimizedXcmsSetObject <- resultPeakpicking$best_settings$xset)}

\item{retcorGroupParameters}{Parameters for retention time correction and optimization}

\item{nSlaves}{Number of slaves the optimization process should spawn.}

\item{plots}{Defines if plots should be generated (TRUE) or not (FALSE) in a
subfolder called "plot_ipo".}
}
\value{
a list with the optimization of parameters for retention time and grouping.
}
\description{
The function optimize parameters considering a set of samples
for the retention time correction and grouping using the IPO Package.
}
\examples{

\dontrun{

file_name <- system.file("extdata", "lcms_opt_xcms_object.rds", package = "NIHSlcms")
optimizedXcmsSetObject <-base::readRDS(file_name)

opt_result_path <- system.file("extdata", "lcms_default_retcorgroup_params.csv", package = "NIHSlcms")
default_retcorgroup_params <- read_IPO_to_XCMS(opt_result_path)


resultRetcorGroup <- lcms_retcorgroup_optimization(optimizedXcmsSetObject,
                                                default_retcorgroup_params,
                                                opt_path = NULL)
print(resultRetcorGroup)
}
}
\concept{optimization functions.}
