% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_retention_time_alignment_plot}
\alias{lcms_retention_time_alignment_plot}
\title{Retention Time Correction Plot}
\usage{
lcms_retention_time_alignment_plot(data)
}
\arguments{
\item{data}{An alignend lcms_dataset.}
}
\value{
The plot for the retention time correction.
}
\description{
It plots the retention time correction vs the original retention time for each of the samples
coloured by sample class.
}
\examples{
\dontrun{
file_name <-  system.file("extdata", "peakdet_align.rds", package = "AlpsLCMS")
data <- base::readRDS(file_name)
rta_plot <- lcms_retention_time_alignment_plot(data)
print(rta_plot)
}
}
\seealso{
Other dataset functions: 
\code{\link{IPO_group_peaks}()},
\code{\link{find_peaks_cwp}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_fill_chrom_peaks}()},
\code{\link{lcms_filter_mz}()},
\code{\link{lcms_filter_polarity}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()},
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_peakpicking_optimization}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_tics}()},
\code{\link{obiAdjust_Rtime}()},
\code{\link{phData}()}

Other retention time correction functions: 
\code{\link{obiAdjust_Rtime}()}

Other visualization functions: 
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_peakpicking_optimization}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_chrom}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_retcorgroup_optimization}()}
}
\concept{dataset functions}
\concept{retention time correction functions}
\concept{visualization functions}
