% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{lcms_spectral_sig_features}
\alias{lcms_spectral_sig_features}
\title{Extract significant features from a MAIT object}
\usage{
lcms_spectral_sig_features(MAIT.object = NULL, pvalue = 0.05,
  p.adj = "none", printCSVfile = FALSE, scale = FALSE,
  parametric = TRUE, var.equal = FALSE, test.fun = NULL,
  jitter = FALSE, jitter.factor = 1, jitter.amount = 0,
  namefun = NULL)
}
\arguments{
\item{MAIT.object}{
A \link{MAIT-class} object where function \link{peakAggregation} has already been applied. The output of the function is going to be an update of the same \link{MAIT-class} object.
}

\item{pvalue}{
P-value threshold. Variables having a p-value lower than this value is considered as a significant variable.
}

\item{p.adj}{
Character with the name of the posthoc method to be applied to correct the pvalues. The supported methods are that of the \link{p.adjust} function
}

\item{printCSVfile}{
Set to TRUE if an output table has to be produced. See function \link{sigPeaksTable} for more information.
}

\item{scale}{
Set to FALSE by default. When set to TRUE, a unit variance scaling of the data when no peak aggregation is performed. If a peak aggregation method is applied, this parameter is ignored.
}

\item{parametric}{
If it is set to TRUE, the statistical tests to be applied will be parametrical tests (e.g. ANOVA, TStudent or Welch's tests). Non-parametrical tests (e.g. Kruskal-Wallis, Mann-Whitney tests) are applied otherwise.
}

\item{var.equal}{
Set to FALSE by default. When set to TRUE, a Student's T-Test is applyied when having 2 classes in the data. If it is set to FALSE, a Welch's test is applyied instead.
}

\item{test.fun}{
Function of the user-defined posthoc method to be applyied.
}

\item{jitter}{
If it is set to TRUE, a jitter noise is added to the data. This is useful when applying Mann-Whitney tests with ties.
}

\item{jitter.factor}{
See argument factor of the function \link{jitter}.
}

\item{jitter.amount}{
See argument amount of the function \link{jitter}.
}

\item{namefun}{
Name of the user-defined posthoc test in the argument test.fun.
}
}
\value{
A MAIT-class object containing the significant features of the scores
slot of MAIT-class object used as an input.
}
\description{
Function lcms_spectral_sig_features takes a MAIT-class object and obtains which of
the variables are significant given a p-value threshold. The parameters of
the significant features can ve printed to an output table (TRUE by default).
Depending on the number of classes in the data, the function chooses between
using ANOVA test or T-Student test.
}
\examples{
file_name <-  system.file("extdata", "peak_table_ann.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name)
peak_table_sig_ann <- lcms_spectral_sig_features(MAIT.object = peak_table,
                                          pvalue=0.05,
                                          p.adj="none",
                                          scale=FALSE)
print(peak_table_sig_ann)
}
\seealso{
Other metabolite identification functions: \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_peak_table_boxplots}},
  \code{\link{lcms_peak_table_pca}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_to_mait}},
  \code{\link{lcms_write_parameter_table}}

Other dataset_peak_table functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_peak_table_boxplots}},
  \code{\link{lcms_peak_table_pca}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_tics}}, \code{\link{lcms_to_mait}}

Other import/export functions: \code{\link{lcms_convert_ipo_to_xcms}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_read_ipo_to_xcms}},
  \code{\link{lcms_read_samples}},
  \code{\link{lcms_rearrange_datafiles_by_class}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_to_mait}},
  \code{\link{lcms_write_opt_params}},
  \code{\link{lcms_write_parameter_table}},
  \code{\link{phData}}
}
\concept{dataset_peak_table functions}
\concept{import/export functions}
\concept{metabolite identification functions}
