% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_xcms_to_mait.R
\name{lcms_to_mait}
\alias{lcms_to_mait}
\title{to MAIT}
\usage{
lcms_to_mait(data_dir = NULL, project_dir = NULL, project = NULL,
  preproc_params = NULL, peak_table = NULL)
}
\arguments{
\item{data_dir}{directory with LC-MS datafiles}

\item{project_dir}{path to the the project directory}

\item{project}{name of the}

\item{preproc_params}{params from XCMS}

\item{peak_table}{XCMS procesed LC-MS data}
}
\value{
A MAIT object
}
\description{
Function to create a MAIT object using data from xcms
}
\details{
\code{MAIT} Package allows to annotate the peaks of the peak table provided by \code{XCMS}.
This is done in three different stages:
\itemize{
\item Annotation with \code{CAMERA}
\item Annotation using predefined biotransformation
\item Annotation using the Human Metabolome Database (HMDB)
}

*Note: The dataset must be converted from an object of the XCMS Package
to an object of the MAIT Package.
}
\examples{
file_name <-  system.file("extdata", "peak_table_imputed.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name)
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

data_dir <-  system.file("extdata", "rearrange_mait", package = "NIHSlcms")
project_dir <-  system.file("extdata", package = "NIHSlcms")
project <- "project"

peak_table_mait <- lcms_to_mait(data_dir = data_dir,
                                project_dir = project_dir,
                                project = project,
                                preproc_params = preproc_params,
                                peak_table = peak_table)
print(peak_table_mait)

}
