% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{lcms_write_opt_params}
\alias{lcms_write_opt_params}
\title{Displaying and Storing optimized settings}
\usage{
lcms_write_opt_params(results_pp, results_rtcg, opt_result_path,
  csv = TRUE, console = TRUE)
}
\arguments{
\item{results_pp}{object from the \code{lcms_peakpicking_optimization} function. If NULL, the default parameters for xcms are loaded.}

\item{results_rtcg}{object from the \code{lcms_corgroup_optimization} function.}

\item{opt_result_path}{A directory to save the parameters file.}

\item{csv}{if TRUE, it writes a file in csv format.}

\item{console}{if TRUE, it displays the params on the console.}
}
\value{
A file with the params from the IPO optimization.
}
\description{
The function allows visulizing the parameter optimization results by \code{IPO} Package
allows in the RStudio console. Also you can save this results in plain text files
(i.e. a .CVS file).
}
\examples{
 opt_result_path <-system.file("extdata", "ipo_opt", package = "NIHSlcms")
 file_name_pp <- system.file("extdata", "result_peakpicking.rds", package = "NIHSlcms")
 file_name_rcg <- system.file("extdata", "result_retcorgroup.rds", package = "NIHSlcms")
 result_peakpicking <- base::readRDS(file_name_pp)
 result_retcorgroup <- base::readRDS(file_name_rcg)

 lcms_write_opt_params(result_peakpicking, result_retcorgroup, opt_result_path)
}
\seealso{
Other optimization functions: \code{\link{lcms_convert_ipo_to_xcms}},
  \code{\link{lcms_default_peakpicking_params}},
  \code{\link{lcms_default_retcorgroup_params}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_read_ipo_to_xcms}}

Other import/export functions: \code{\link{lcms_convert_ipo_to_xcms}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_read_ipo_to_xcms}},
  \code{\link{lcms_read_samples}},
  \code{\link{lcms_rearrange_datafiles_by_class}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_spectral_sig_features}},
  \code{\link{lcms_to_mait}},
  \code{\link{lcms_write_parameter_table}},
  \code{\link{phData}}
}
\concept{import/export functions}
\concept{optimization functions}
