% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_xcms_to_mait.R
\name{lcms_write_parameter_table}
\alias{lcms_write_parameter_table}
\title{Write Parameters Table for MAIT functionalities}
\usage{
lcms_write_parameter_table(listParameters, folder)
}
\arguments{
\item{listParameters}{Parameters}

\item{folder}{A directory to store a .csv file where the  parameter table is stored.}
}
\value{
A template with MAIT parameters.
}
\description{
Write Parameters Table for MAIT functionalities
}
\examples{
data_dir <-  system.file("extdata", "rearrange_mait", package = "NIHSlcms")
mait_params_path <- system.file("extdata", "results_project", package = "NIHSlcms")
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)
parameters <- c(dataDir = data_dir, preproc_params)

quiet <- function(x) {
             base::sink(base::tempfile())
             base::on.exit(base::sink())
             base::invisible(base::force(x))}

quiet(getClassDef("MAIT","MAIT"))
MAIT.object = signature(MAIT.Object = "MAIT")
MAIT.object <- methods::new("MAIT")
MAIT.object@RawData@parameters@sampleProcessing <- parameters
lcms_write_parameter_table(MAIT::parameters(MAIT.object), folder = mait_params_path)
}
