% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_converter_lister.R
\name{list_mzxml_samples}
\alias{list_mzxml_samples}
\title{RAW converter}
\usage{
list_mzxml_samples(sample_path, file_format = "raw", rawconverter)
}
\arguments{
\item{sample_path}{Directory in which the samples are.}

\item{file_format}{Format of the LC-MS files (e.g. file_format = "raw").}

\item{rawconverter}{Directory in where the RawConverter application is located.}
}
\value{
a list of the LC-MS files in a readable format to create the \code{lcms_dataset}.
}
\description{
The function lists and converts samples in ".raw" format to ".mzXML" format.
It is the first step to create a \code{lcms_dataset}. It requires the previous
download and installation of the \emph{RawConverter} application.
}
\examples{
\dontrun{
rawconverter_path <-"C:/Users/Biosignal/Desktop/Nestle_2019/LCMS"
path <- system.file("extdata", "1.mzXML", package = "NIHSlcms")
samples_mzxml <- list_mzxml_samples(path,
                                    file_format = "mzXML",
                                    rawconverter = rawconverter_path)
}

\dontrun{
path <- system.file("extdata", "1.raw", package = "NIHSlcms")
samples_mzxml <- list_mzxml_samples(path,
                                    file_format = "raw",
                                    rawconverter = rawconverter_path)
}
samples_mzxml <- list_mzxml_samples(path,
                                    file_format = "mzXML",
                                   rawconverter = rawconverter_path)
}
