% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lmcs_plot_chrom_peak_image}
\alias{lmcs_plot_chrom_peak_image}
\title{Image of Chromatographic Peaks by sample}
\usage{
lmcs_plot_chrom_peak_image(x, binSize = 30, xlim = NULL, log = FALSE,
  xlab = "retention time", yaxt = par("yaxt"),
  main = "Chromatographic peak counts", ...)
}
\arguments{
\item{dataset}{An dataset}
}
\value{
An image with the detected chromatographic peak, for each sample
}
\description{
It plots the an image of the chromatographic peaks for each sample. This function is useful if
you are interested in knowing the effect of the retention time correction on the chromatographic axis.
}
\examples{
file_name <-  system.file("extdata", "peakdet_align.rds", package = "NIHSlcms")
dataset <- base::readRDS(file_name)
chr_peak_image <- lmcs_plot_chrom_peak_image(dataset, binSize = 5, xlim = NULL, log = FALSE,
                                               xlab = "retention time (min)", yaxt = par("yaxt"))
                                               title(main ="Detected Peaks (Aligned)")
}
