% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_load_save.R
\name{load_and_save_functions}
\alias{load_and_save_functions}
\alias{lcms_dataset_load}
\alias{lcms_dataset_save}
\title{Functions to load and save lcms_dataset objects}
\usage{
lcms_dataset_load(file_name)

lcms_dataset_save(lcms_dataset, file_name, ...)
}
\arguments{
\item{file_name}{The file name to load or save to}

\item{lcms_dataset}{An object from the [lcms_dataset_family]}

\item{...}{Additional arguments passed to [saveRDS].}
}
\description{
Functions to load and save lcms_dataset objects
}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_meta_export}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_meta_export}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other import/export functions: \code{\link{lcms_meta_export}}
}
\concept{import/export functions}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
