% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{models_stability_plot_bootstrap}
\alias{models_stability_plot_bootstrap}
\title{Models stability plot}
\usage{
models_stability_plot_bootstrap(bp_results)
}
\arguments{
\item{bp_results}{bp_kfold_VIP_analysis results}
}
\value{
A plot of models stability
}
\description{
Plot stability among models of the external cross validation
}
