% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{models_stability_plot_plsda}
\alias{models_stability_plot_plsda}
\title{Models stability plot}
\usage{
models_stability_plot_plsda(model)
}
\arguments{
\item{model}{A lcms_data_analysis_model}
}
\value{
A plot of models stability
}
\description{
Plot stability among models of the external cross validation
}
