% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHSlcms.R
\docType{package}
\name{NIHSlcms-package}
\alias{NIHSlcms}
\alias{NIHSlcms-package}
\title{NIHSlcms: Signal processing of LC-MS metabolomics}
\description{
NIHSlcms allows ... metabolomics.
}
\examples{
library(NIHSlcms)
}
\author{
\strong{Maintainer}: Luis Fernandez \email{lfernandez@ibecbarcelona.eu}

Authors:
\itemize{
  \item Francisco Madrid-Gambin \email{fmadrid@ibecbarcelona.eu}
  \item Sergio Oller \email{soller@ibecbarcelona.eu}
  \item Ivan Montoliu Roura \email{Ivan.MontoliuRoura@rd.nestle.com}
}

Other contributors:
\itemize{
  \item Nestlé Institute of Health Sciences [copyright holder]
  \item Institute for Bioengineering of Catalonia [copyright holder]
}

}
