% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{obiAdjust_Rtime}
\alias{obiAdjust_Rtime}
\title{Retention Time Correction}
\usage{
obiAdjust_Rtime(peakdet, params)
}
\arguments{
\item{peakdet}{A lcms_dataset with detected peaks obtained from the
\code{find_peaks_cwp} function.}

\item{params}{A converted parameters template from IPO parameters.}
}
\value{
A lcms_dataset with (1) detected (\code{find_peaks_cwp} function).
and (2) aligned (\code{obiAdjust_Rtime} function) peaks
}
\description{
Retention time correction is performed using \emph{'obiwarp'} method.
}
\details{
Note: signal processing generally  consists in three main steps:
(1) peak detection (\code{find_peaks_cwp} function),
(2) peak alignment (\code{obiAdjust_Rtime} function) and
(3) peak correspondence (\code{IPO_group_peaks} function).
The optimized set of parameters for this signal preprocessing can be obatained from \code{IPO} Package.
}
\examples{
\dontrun{
file_name <-  system.file("extdata", "peakdet.rds", package = "AlpsLCMS")
peakdet <- base::readRDS(file_name)
opt_result_path <-  system.file("extdata", package = "AlpsLCMS")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

peakdet_align <- obiAdjust_Rtime(peakdet, params = preproc_params)
print(peakdet_align)
}
}
\seealso{
Other dataset functions: 
\code{\link{IPO_group_peaks}()},
\code{\link{find_peaks_cwp}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_fill_chrom_peaks}()},
\code{\link{lcms_filter_mz}()},
\code{\link{lcms_filter_polarity}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()},
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_retention_time_alignment_plot}()},
\code{\link{lcms_tics}()},
\code{\link{peakpicking_optimization}()},
\code{\link{phData}()}

Other retention time correction functions: 
\code{\link{lcms_retention_time_alignment_plot}()}
}
\concept{dataset functions}
\concept{retention time correction functions}
