% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{PeakDensityPar}
\alias{PeakDensityPar}
\title{Peak density parameters}
\usage{
PeakDensityPar(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[xcms:groupChromPeaks-density]{xcms::PeakDensityParam}}
  \describe{
    \item{\code{sampleGroups}}{A vector of the same length than samples defining the
sample group assignments (i.e. which samples belong to which sample
group). This parameter is mandatory for the \code{PeakDensityParam}
and has to be provided also if there is no sample grouping in the
experiment (in which case all samples should be assigned to the
same group).}
    \item{\code{bw}}{\code{numeric(1)} defining the bandwidth (standard deviation ot the
smoothing kernel) to be used. This argument is passed to the
[density() method.}
    \item{\code{minFraction}}{\code{numeric(1)} defining the minimum fraction of samples
in at least one sample group in which the peaks have to be present to be
considered as a peak group (feature).}
    \item{\code{minSamples}}{\code{numeric(1)} with the minimum number of samples in at
least one sample group in which the peaks have to be detected to be
considered a peak group (feature).}
    \item{\code{binSize}}{\code{numeric(1)} defining the size of the overlapping slices
in mz dimension.}
    \item{\code{maxFeatures}}{\code{numeric(1)} with the maximum number of peak groups
to be identified in a single mz slice.}
  }}
}
\value{
PeakDensityParam S4 object
}
\description{
The method creates a template with the parameters for \strong{correspondence}.
Extracted from the function \code{xcms::PeakDensityParam}.
}
\section{Slots}{


\describe{
\item{\code{.__classVersion__,sampleGroups,bw,minFraction,minSamples,binSize,maxFeatures}}{See corresponding parameter above. \code{.__classVersion__} stores
the version from the class. Slots values should exclusively be accessed
\emph{via} the corresponding getter and setter methods listed above.}
}
}

\examples{
## Create a PeakDensityParam object
p <- PeakDensityParam(binSize = 0.05, sampleGroups = c(1, 1, 2, 2))
## Change hte minSamples slot
minSamples(p) <- 3
p

##############################
## Chromatographic peak detection and grouping.
##
## Below we perform first a peak detection (using the matchedFilter
## method) on some of the test files from the faahKO package followed by
## a peak grouping using the density method.
library(faahKO)
library(MSnbase)
fls <- dir(system.file("cdf/KO", package = "faahKO"), recursive = TRUE,
           full.names = TRUE)

## Reading 2 of the KO samples
raw_data <- readMSData(fls[1:2], mode = "onDisk")

## Perform the chromatographic peak detection using the matchedFilter method.
mfp <- MatchedFilterParam(snthresh = 20, binSize = 1)
res <- findChromPeaks(raw_data, param = mfp)

head(chromPeaks(res))
## The number of peaks identified per sample:
table(chromPeaks(res)[, "sample"])

## Performing the chromatographic peak grouping. Assigning all samples to
## the same sample group.
fdp <- PeakDensityParam(sampleGroups = rep(1, length(fileNames(res))))
res <- groupChromPeaks(res, fdp)

## The definition of the features (peak groups):
featureDefinitions(res)

## Using the featureValues method to extract a matrix with the
## intensities of the features per sample.
head(featureValues(res, value = "into"))

## The process history:
processHistory(res)
}
\references{
Colin A. Smith, Elizabeth J. Want, Grace O'Maille, Ruben Abagyan and
Gary Siuzdak. "XCMS: Processing Mass Spectrometry Data for Metabolite
Profiling Using Nonlinear Peak Alignment, Matching, and Identification"
Anal. Chem. 2006, 78:779-787.
}
\seealso{
The \code{\link[xcms:do_groupChromPeaks_density]{do_groupChromPeaks_density()}} core API function and \code{\link[xcms:group.density]{group.density()}}
for the old user interface.

\code{\link[xcms:plotChromPeakDensity]{plotChromPeakDensity()}} to plot peak densities and evaluate different
algorithm settings.

\code{\link[xcms:XCMSnExp-class]{featureDefinitions()}} and \code{\link[xcms:XCMSnExp-peak-grouping-results]{featureValues()}} for methods to access the
features (i.e. the peak grouping results).

\link[xcms]{XCMSnExp} for the object containing the results of the correspondence.

\code{\link[xcms:plotChromPeakDensity]{plotChromPeakDensity()}} for plotting chromatographic peak density with the
possibility to test different parameter settings.

Other peak grouping methods: \code{\link[xcms]{groupChromPeaks-mzClust}},
  \code{\link[xcms]{groupChromPeaks-nearest}},
  \code{\link[xcms]{groupChromPeaks}}
}
\author{
Colin Smith, Johannes Rainer
}
