% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{PeakGroupsPar}
\alias{PeakGroupsPar}
\title{Peak density parameters}
\usage{
PeakGroupsPar(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[xcms:adjustRtime-peakGroups]{xcms::PeakGroupsParam}}
  \describe{
    \item{\code{minFraction}}{\code{numeric(1)} between 0 and 1 defining the minimum
required fraction of samples in which peaks for the peak group were
identified. Peak groups passing this criteria will aligned across
samples and retention times of individual spectra will be adjusted
based on this alignment. For \code{minFraction = 1} the peak group
has to contain peaks in all samples of the experiment. Note that if
\code{subset} is provided, the specified fraction is relative to the
defined subset of samples and not to the total number of samples within
the experiment (i.e. a peak has to be present in the specified
proportion of subset samples).}
    \item{\code{extraPeaks}}{\code{numeric(1)} defining the maximal number of
additional peaks for all samples to be assigned to a peak group (i.e.
feature) for retention time correction. For a data set with 6 samples,
\code{extraPeaks = 1} uses all peak groups with a total peak count
\code{<= 6 + 1}. The total peak count is the total number of peaks being
assigned to a peak group and considers also multiple peaks within a
sample being assigned to the group.}
    \item{\code{smooth}}{character defining the function to be used, to interpolate
corrected retention times for all peak groups. Either \code{"loess"} or
\code{"linear"}.}
    \item{\code{span}}{\code{numeric(1)} defining the degree of smoothing (if
\code{smooth = "loess"}). This parameter is passed to the internal call
to \code{\link{loess}}.}
    \item{\code{family}}{character defining the method to be used for loess smoothing.
Allowed values are \code{"gaussian"} and \code{"symmetric"}.See
\code{\link{loess}} for more information.}
    \item{\code{peakGroupsMatrix}}{optional \code{matrix} of (raw) retention times for
the peak groups on which the alignment should be performed. Each column
represents a sample, each row a feature/peak group. Such a matrix is
for example returned by the \code{\link[xcms]{adjustRtimePeakGroups}} method.}
    \item{\code{subset}}{\code{integer} with the indices of samples within the
experiment on which the alignment models should be estimated. Samples
not part of the subset are adjusted based on the closest subset sample.
See description above for more details.}
    \item{\code{subsetAdjust}}{\code{character} specifying the method with which
non-subset samples should be adjusted. Supported options are
\code{"previous"} and \code{"average"} (default). See description above
for more information.}
  }}
}
\value{
PeakGroupsParam S4 object
}
\description{
The method creates a template with the parameters for \strong{alignment}.
Extracted from the function \code{xcms::PeakGroupsParam}.
}
