% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{permutation_test_model}
\alias{permutation_test_model}
\title{Permutation test}
\usage{
permutation_test_model(
  dataset,
  y_column,
  identity_column,
  balance_in_train = NULL,
  external_val,
  internal_val,
  data_analysis_method,
  nPerm = 50
)
}
\arguments{
\item{dataset}{An \link{lcms_dataset_family} object}

\item{y_column}{A string with the name of the y column (present in the
metadata of the dataset)}

\item{identity_column}{\code{NULL} or a string with the name of the identity column (present in the
metadata of the dataset).}

\item{external_val}{A list with two elements: \code{iterations} and \code{test_size}.
See \link{random_subsampling} for further details}

\item{internal_val}{A list with two elements: \code{iterations} and \code{test_size}.
See \link{random_subsampling} for further details}

\item{data_analysis_method}{An \link{lcms_data_analysis_method} object}

\item{nPerm}{number of permutations}
}
\value{
A permutation matrix with permuted values
}
\description{
Make permutations with data and default settings from an lcms_data_analysis_method
}
