% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{phData}
\alias{phData}
\title{Generic function to access the phenotypic data (Biobase)}
\usage{
phData(object)
}
\arguments{
\item{object}{Object, possibly derived from \code{\link{eSet-class}}
    or \code{\linkS4class{AnnotatedDataFrame}}.}

\item{An}{object from the \link{lcms_dataset_family}}
}
\value{
phenoData returns an object containing information on
both variable values and variable meta-data. varLabels returns
a character vector of measured variables. pData returns a data
frame with samples as rows, variables as columns. varMetadata
returns a data frame with variable names as rows, description
tags (e.g., unit of measurement) as columns.
}
\description{
It uses \link[Biobase:pData]{Biobase::pData} to access phenotypic data such as
metadata, covariates, etc...
}
\examples{
dataset <- lcms_dataset_load(system.file("extdata",
                                         "dataset.rds",
                                          package = "NIHSlcms"))
xlsx_file <- system.file("extdata", "exported_metadata.xlsx", package = "NIHSlcms")
metadata<- readxl::read_excel(xlsx_file)
dataset_metadata <- lcms_meta_add(dataset, metadata, by = "sampleNames")

head(phData(dataset_metadata))[1:4]
}
\seealso{
Other metadata functions: \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}}

Other dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_export}}

Other import/export functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_read_samples}}
}
\concept{dataset functions}
\concept{import/export functions}
\concept{metadata functions}
