% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{plot_bootstrap_multimodel}
\alias{plot_bootstrap_multimodel}
\title{Bootstrap plot predictions}
\usage{
plot_bootstrap_multimodel(bp_results, dataset, y_column, plot = TRUE)
}
\arguments{
\item{bp_results}{bp_kfold_VIP_analysis results}

\item{dataset}{An \link{lcms_dataset_family} object}

\item{y_column}{A string with the name of the y column (present in the
metadata of the dataset)}

\item{plot}{A boolean that indicate if results are plotted or not}
}
\value{
A plot of the results or a ggplot object
}
\description{
Bootstrap plot predictions
}
