% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plot_plsda_multimodel}
\alias{plot_plsda_multimodel}
\title{Multi PLDSA model plot predictions}
\usage{
plot_plsda_multimodel(model, plot = TRUE)
}
\arguments{
\item{model}{A lcms_data_analysis_model}

\item{plot}{A boolean that indicate if results are plotted or not}
}
\value{
A plot of the results or a ggplot object
}
\description{
Multi PLDSA model plot predictions
}
