% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plot_plsda_samples}
\alias{plot_plsda_samples}
\title{Plot PLSDA predictions}
\usage{
plot_plsda_samples(model, newdata = NULL, plot = TRUE)
}
\arguments{
\item{model}{A plsda model}

\item{newdata}{newdata to predict, if not included model$X_test will be used}

\item{plot}{A boolean that indicate if results are plotted or not}
}
\value{
A plot of the samples or a ggplot object
}
\description{
Plot PLSDA predictions
}
