% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{plot_vip_scores}
\alias{plot_vip_scores}
\title{Plot vip scores of bootstrap}
\usage{
plot_vip_scores(vip_means, error, nbootstrap, plot = TRUE)
}
\arguments{
\item{vip_means}{vips means values of bootstraps}

\item{error}{error tolerated, calculated in the bootstrap}

\item{nbootstrap}{number of bootstraps realiced}

\item{plot}{A boolean that indicate if results are plotted or not}
}
\value{
A plot of the results or a ggplot object
}
\description{
Plot vip scores of bootstrap
}
