% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{raw_data}
\alias{raw_data}
\title{Raw data extractor from a MAIT object}
\usage{
raw_data(MAIT.object)
}
\arguments{
\item{MAIT.object}{A MAIT-class object}
}
\value{
A list containing either a xcmsSet or a xsAnnotate object.
}
\description{
Function rawData extracts the raw data used to build the MAIT-class object
}
\examples{
\dontrun{
peak_table_MAIT <- to_MAIT(dataDir = dataDir,
        project = project,
        preproc_params = preproc_params,
        peakTable = peakTable)

peak_table_ANN <- peak_annotation(MAIT.object = peak_table_MAIT)
raw_data(peak_table_ANN)
}
}
