% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPO_to_XCMS.R
\name{read_IPO_to_XCMS}
\alias{read_IPO_to_XCMS}
\title{Read IPO parameters to XCMS format}
\usage{
read_IPO_to_XCMS(opt_result_path)
}
\arguments{
\item{opt_result_path}{A directory to save the parameters file}
}
\value{
A display of the chosen parameters
}
\description{
The function reads and converts from \code{IPO} to \code{XCMS} variable formats: The variable names in IPO and XCMS
presented some mismatches, thus we had to create a function called \emph{IPO_to_XCMS}
to achieve compatibility between packages.
}
\examples{
write_opt_params(resultPeakpicking, resultRetcorGroup, my_directory)
}
