% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files_by_class.R
\name{rearrange_datafiles_by_class}
\alias{rearrange_datafiles_by_class}
\title{Title}
\usage{
rearrange_datafiles_by_class(lcms_dataset, dataDir)
}
\arguments{
\item{lcms_dataset}{An [lcms_dataset_family] object}
}
\description{
Title
}
