% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/utils_future.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{plan}
\alias{sequential}
\alias{multicore}
\alias{multiprocess}
\alias{multisession}
\alias{cluster}
\alias{remote}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{future}{\code{\link[future]{cluster}}, \code{\link[future]{multicore}}, \code{\link[future]{multiprocess}}, \code{\link[future]{multisession}}, \code{\link[future]{plan}}, \code{\link[future]{remote}}, \code{\link[future]{sequential}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

