% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_MAIT.R
\name{to_MAIT}
\alias{to_MAIT}
\title{to MAIT}
\usage{
to_MAIT(dataDir = NULL, project = NULL, preproc_params = NULL,
  peakTable = NULL)
}
\arguments{
\item{dataDir}{directory with LC-MS datafiles}

\item{project}{name of the project}

\item{preproc_params}{params from XCMS}

\item{peakTable}{XCMS procesed LC-MS data}
}
\value{
A MAIT object
}
\description{
Function to create a MAIT object
}
