% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_MAIT.R
\name{writeParameterTable}
\alias{writeParameterTable}
\title{Write Parameters Table for MAIT functionalities}
\usage{
writeParameterTable(listParameters, folder)
}
\arguments{
\item{listParameters}{Parameters}

\item{folder}{A directory}
}
\value{
A template with MAIT parameters. Also write the `MAITparameters.csv´ file.
}
\description{
\code{MAIT} Package allows to annotate the peaks of the peak table provided by \code{XCMS}.
This is done in three different stages:
\itemize{
\item Annotation with \code{CAMERA}
\item Annotation using predefined biotransformation
\item Annotation using the Human Metabolome Database (HMDB)
}
}
\details{
*Note: The dataset must be converted from an object of the XCMS Package
to an object of the MAIT Package.
}
