% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{adjustRT}
\alias{adjustRT}
\title{Peak alignment}
\usage{
adjustRT(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[xcms:adjustRtime]{xcms::adjustRtime}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
XCMSnExp object
}
\description{
\code{adjustRT} is a wrapper of the \code{xcms::adjustRtime} from \code{xcms} package. It
performs the \strong{alignment} (grouping) of chromatographic peaks of the
samples.
}
\seealso{
\code{\link[xcms]{retcor}} for the \emph{old} retention time correction
    methods.
    \code{\link[xcms]{plotAdjustedRtime}} for visualization of alignment results.

Other retention time correction methods: \code{\link[xcms]{adjustRtime-obiwarp}},
  \code{\link[xcms]{adjustRtime-peakGroups}}
}
\author{
Johannes Rainer
}
