% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retention_Time_Correction.R
\name{align_Rtime}
\alias{align_Rtime}
\title{Retention Time Correction}
\usage{
align_Rtime(peakdet, params)
}
\arguments{
\item{peakdet}{A lcms_dataset with detected peaks from the
\code{findChromPeaks_cwp} function}

\item{params}{A converted parameters template from IPO parameters.}
}
\value{
A lcms_dataset with (1) detected (\code{findChromPeaks_cwp} function)
and (2) aligned (\code{align_Rtime} function) peaks
}
\description{
Retention time correction is performed  using \emph{'obiwarp'} method.
Its optimum parameters are obtaine from \code{IPO} Package.
Note: signal processing generally  consists in three main steps:
(1) peak detection (\code{findChromPeaks_cwp} function),
(2) peak alignment (\code{align_Rtime} function) and
(3) peak correspondence (\code{group_peaks} function).
}
\examples{
params <- convert_IPO_to_XCMS(IPO_params)
peakdet <- findChromPeaks_cwp(dataset, params = params)
peakdet_align <- align_Rtime(peakdet, params = params)

}
