% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{cromplot}
\alias{cromplot}
\title{cromplot}
\usage{
cromplot(dataset, rt, mz, metabolite = "metabolite")
}
\arguments{
\item{dataset}{a lcms_dataset}

\item{rt}{known retention time of a metabolite that surely is in the dataset}

\item{mz}{known mass of a metabolite that surely is in the dataset}

\item{metabolite}{character. The name of the metabolite (for plotting)}
}
\value{

}
\description{
Function that creates the plots of a know metabolite to observe and set the
parameters for xcms preprocessing. Parameters such as peak width, noise
level, or potential neighbors can be studied.
}
