% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{feature_reduction}
\alias{feature_reduction}
\title{Feature reduction}
\usage{
feature_reduction(mdataImputed, Representative_ions, RC)
}
\arguments{
\item{mdataImputed}{data matrix with samples in rows and features in columns}

\item{Representative_ions}{data frame from the list given by \link{labelled_adducts} function}

\item{RC}{hclust object from the \link{do.findmain} function}
}
\value{
A data frame with important features based on the clustering
}
\description{
Function to perform feature reduction from the \link{clustering}.
}
