% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{feature_reduction}
\alias{feature_reduction}
\title{Feature reduction}
\usage{
feature_reduction(RC,
xdata,
loaded = TRUE)
}
\arguments{
\item{RC}{hclust object from the \link{do_findmain} function}

\item{xdataImp}{data matrix with samples in rows and features in columns. The
matrix should not include metadata or any columns without LC-MS data. Note
that depending on the origins of the dataset, colnames might have a
character starting with "X" avoiding a number in the first character (see
\code{loaded = TRUE})}

\item{loaded}{if matrix is loaded into R, an "X" will be added to each
colname. In this case, set \code{loaded = TRUE}.}
}
\value{
A list containing:
$singletons: data frame with features that has not been clustered.
$xdata_mol_ions: data frame with the representative ions of each cluster
$xdata_reduced: data frame with the reduced dataset (singletons +
representative ions of each cluster)
}
\description{
\code{feature_reduction} is a function to reduce the feature table using the info
from the \link{clustering}. A representative feature/ion is selected from each
cluster discarding the rest of the feature of the same cluster (normally, the
most intense one).
}
