% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getRamSt}
\alias{getRamSt}
\title{Get st parameter value for clustering}
\usage{
getRamSt(lcms_dataset)
}
\arguments{
\item{lcms_dataset}{A lcms_dataset}
}
\value{
the recommended value for the \code{st} parameter in \link{clustering}
}
\description{
\code{getRamSt} function calculates the optimal value for the parameter \code{st} the \link{clustering}
function. It uses the \code{lcms_dataset} after the xcms preprocessing.
}
