% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{labelled_adducts}
\alias{labelled_adducts}
\title{Labelled adducts}
\usage{
labelled_adducts(RC)
}
\arguments{
\item{RC}{hclust object from the \link{do.findmain} function}
}
\value{

}
\description{
Function ton extract labelled adducts from \link{clustering}.
}
