% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_align_rtime}
\alias{lcms_align_rtime}
\title{Retention Time Correction}
\usage{
lcms_align_rtime(peakdet, params)
}
\arguments{
\item{peakdet}{A dataset with detected peaks from the
\code{lcms_find_chrom_peaks_cwp} function}

\item{params}{A converted parameters template from IPO parameters.}
}
\value{
A dataset with (1) detected (\code{lcms_find_chrom_peaks_cwp} function)
and (2) aligned (\code{lcms_align_rtime} function) peaks
}
\description{
Retention time correction is performed using \emph{'obiwarp'} method.
Its optimum parameters are obtaine from \code{IPO} Package.
Note: signal processing generally  consists in three main steps:
(1) peak detection (\code{lcms_find_chrom_peaks_cwp} function),
(2) peak alignment (\code{lcms_align_rtime} function) and
(3) peak correspondence (\code{lcms_group_peaks} function).
}
\examples{
file_name <-  system.file("extdata", "peakdet.rds", package = "NIHSlcms")
peakdet <- base::readRDS(file_name)
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

peakdet_align <- lcms_align_rtime(peakdet, params = preproc_params)
print(peakdet_align)

}
