% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{lcms_convert_ipo_to_xcms}
\alias{lcms_convert_ipo_to_xcms}
\title{Converts IPO parameters to XCMS format}
\usage{
lcms_convert_ipo_to_xcms(params)
}
\arguments{
\item{params}{A list with the IPO parameters, generated with the \code{lcms_write_opt_params} function.}
}
\value{
Parameters in XCMS format
}
\description{
The function converts from \code{IPO} to \code{XCMS} variable formats: The variable names in IPO and XCMS
presented some mismatches, thus the function \emph{lcms_convert_ipo_to_xcms} was created
to achieve compatibility between packages when these parameters are stored in a list.
}
\examples{
\dontrun{
opt_result_path <-  system.file("extdata","ipo_opt", "params.rds", package = "AlpsLCMS")
params <- base::readRDS(opt_result_path)
preproc_params <- lcms_convert_ipo_to_xcms(params)
print(preproc_params)
}
}
\seealso{
Other optimization functions: 
\code{\link{default_peakpicking_params}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_retcorgroup_optimization}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{peakpicking_optimization}()}

Other import/export functions: 
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()},
\code{\link{phData}()}
}
\concept{import/export functions}
\concept{optimization functions}
