% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{lcms_corgroup_optimization}
\alias{lcms_corgroup_optimization}
\title{Optimization of retention time correction and grouping parameters.}
\usage{
lcms_corgroup_optimization(optimizedXcmsSetObject, retcorGroupParameters,
  nSlaves = 4, plots = TRUE)
}
\arguments{
\item{optimizedXcmsSetObject}{XCMS object conatining the `best_settings` parameters.
This object may be created after running `lcms_peakpicking_optimization` and extract
best_settings$xset (e.g. optimizedXcmsSetObject <- resultPeakpicking$best_settings$xset)}

\item{retcorGroupParameters}{Parameters for retention time correction and optimization}

\item{nSlaves}{Number of slaves the optimization process should spawn.}

\item{plots}{Defines if plots should be generated (TRUE) or not (FALSE) in a
subfolder called "plot_ipo".}
}
\value{
a list with the optimization of parameters for retention time and grouping.
}
\description{
The function optimize parameters considering a set of samples
for the retention time correction and grouping using the IPO Package.
}
\seealso{
Other optimization functions: \code{\link{lcms_default_corgroup_params}},
  \code{\link{lcms_default_peakpeaking_params}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{write_opt_params}}
}
\concept{optimization functions}
