% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{lcms_dataset_load}
\alias{lcms_dataset_load}
\title{Load lcms_datasets}
\usage{
lcms_dataset_load(file_name)
}
\arguments{
\item{file_name}{The file name to load.}
}
\value{
A lcms_dataset.
}
\description{
Function to load a lcms_dataset object.
}
\examples{
\dontrun{
file_name <- system.file("extdata", "dataset.rds", package = "AlpsLCMS")
dataset <- lcms_dataset_load(file_name)
print(dataset)
}
}
\seealso{
Other dataset functions: 
\code{\link{IPO_group_peaks}()},
\code{\link{find_peaks_cwp}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_fill_chrom_peaks}()},
\code{\link{lcms_filter_mz}()},
\code{\link{lcms_filter_polarity}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()},
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_retention_time_alignment_plot}()},
\code{\link{lcms_tics}()},
\code{\link{obiAdjust_Rtime}()},
\code{\link{peakpicking_optimization}()},
\code{\link{phData}()}

Other dataset_peak_table functions: 
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_tics}()},
\code{\link{lcms_to_mait}()}

Other import/export functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()},
\code{\link{phData}()}
}
\concept{dataset functions}
\concept{dataset_peak_table functions}
\concept{import/export functions}
