% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_load_save.R
\name{lcms_dataset_load}
\alias{lcms_dataset_load}
\title{Function to load a lcms_dataset object}
\usage{
lcms_dataset_load(file_name)
}
\arguments{
\item{file_name}{The file name to load}
}
\value{
An object from the \link{lcms_dataset_family}
}
\description{
Function to load a lcms_dataset object
}
\examples{
file_name <- system.file("extdata", "lcms_dataset.rds", package = "NIHSlcms")
lcms_dataset <- lcms_dataset_load(file_name)
print(lcms_dataset)

}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_filterPolarity}},
  \code{\link{lcms_filterRTmin}},
  \code{\link{lcms_filterSampleType}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other import/export functions: \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}}
}
\concept{import/export functions}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
