% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_load_save_2.R
\name{lcms_dataset_save}
\alias{lcms_dataset_save}
\title{Function to save a lcms_dataset object}
\usage{
lcms_dataset_save(lcms_dataset, file_name, ...)
}
\arguments{
\item{lcms_dataset}{An object from the \link{lcms_dataset_family}}

\item{file_name}{The file name to save to}

\item{...}{Additional arguments passed to \link{saveRDS}.}
}
\description{
Function to save a lcms_dataset object
}
