% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_parameter_optimization.R
\name{lcms_default_peakpicking_params}
\alias{lcms_default_peakpicking_params}
\title{Default parameters for peak picking optimization}
\usage{
lcms_default_peakpicking_params(noise = 5000, snthresh = 10,
  min_peakwidth = c(10, 30), max_peakwidth = c(35, 120),
  optimize = TRUE)
}
\arguments{
\item{noise}{numeric, minimum intensity needed to be included in the analysis.}

\item{snthresh}{numeric, set the signal to noise ratio threshold to be included.}

\item{min_peakwidth}{numeric(two values) with the expected minimal peak width in
chromatographic dimension. Set as a range (min, max) in seconds.}

\item{max_peakwidth}{numeric(two values) with the expected maximum peak width in
chromatographic dimension. Set as a range (min, max) in seconds.}

\item{optimize}{by default is TRUE. If FALSE, the function does not initialze the parameters
the parameters}
}
\value{
A parameters template for peak picking optimization
}
\description{
The function creates default parameters for optimizing
peakpicking algorithms.We perform parameter optimization
on the XCMS preprocessing algorithms using the IPO Package.
This includes Peak Detection (‘Centwave’ and ‘Matched Filter’),
Retention Time Correction (‘obiwarp’) and Peak Correspondence
(‘Density’).
Use this function to generate the template within the
\code{lcms_peakpicking_optimization} function.
}
\examples{
default_peakpicking_params <- lcms_default_peakpicking_params(optimize = TRUE)
print(default_peakpicking_params)

}
\seealso{
Other optimization functions: \code{\link{lcms_default_retcorgroup_params}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_write_opt_params}}
}
\concept{optimization functions}
