% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_filter_mz}
\alias{lcms_filter_mz}
\title{Filter by mass/charge}
\usage{
lcms_filter_mz(dataset, mz)
}
\arguments{
\item{dataset}{An \link{lcms_dataset_family} object}

\item{mz}{The range of masses to filter}
}
\value{
A filtered \link{lcms_dataset_family} object with the selecter m/z range
}
\description{
The function filters the dataset by m/z
}
\examples{
mz_range = c(300, 400)
dataset <- lcms_dataset_load(system.file("extdata","dataset_pos.rds",package = "NIHSlcms"))
dataset_pos_mz <-lcms_filter_mz(dataset, mz = mz_range)

print(dataset_pos_mz)
}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_filtering functions: \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}}
}
\concept{lcms_dataset functions}
\concept{lcms_filtering functions}
