% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_filter_polarity}
\alias{lcms_filter_polarity}
\title{Filter an experiment by its polarity}
\usage{
lcms_filter_polarity(lcms_dataset, polarity.)
}
\arguments{
\item{lcms_dataset}{A MSnExp object.}

\item{polarity.}{The polarity to keep.}
}
\value{
A filtered lcms_dataset with the selected polarity.
}
\description{
This function was merged in
the MSnbase package on 2018-12-10 as proposed at:
}
\details{
https://github.com/lgatto/MSnbase/issues/388

Once MSnbase is released (Bioconductor 3.9, maybe once R 3.6 is out),
remove this function.
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_read_samples("yourDir/sample1.mzXML", mode = "onDisk")
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

is.positive(dataset_pos)
}
}
\seealso{
Other dataset functions: 
\code{\link{IPO_group_peaks}()},
\code{\link{find_peaks_cwp}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_fill_chrom_peaks}()},
\code{\link{lcms_filter_mz}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()},
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_retention_time_alignment_plot}()},
\code{\link{lcms_tics}()},
\code{\link{obiAdjust_Rtime}()},
\code{\link{peakpicking_optimization}()},
\code{\link{phData}()}

Other filtering functions: 
\code{\link{lcms_filter_mz}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()}
}
\concept{dataset functions}
\concept{filtering functions}
