% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{lcms_identify_metabolites}
\alias{lcms_identify_metabolites}
\title{Metabolite search}
\usage{
lcms_identify_metabolites(
  MAIT.object = NULL,
  peakTolerance = 0.005,
  database = NULL,
  polarity = "positive",
  printCSVfile = TRUE
)
}
\arguments{
\item{MAIT.object}{
A \link[MAIT]{MAIT-class} object where significant features have already been found.
}

\item{peakTolerance}{
Maximum difference between the peak masses differences and the values shown in the database to be considered as a match. As default the value is 0.005 Da.
}

\item{database}{
User-defined input table. If it is set to NULL, the default \pkg{MAIT} database is selected to perform the metabolite identification.
}

\item{polarity}{
Character parameter that can be set to "positive" or "negative" depending on the polarity in which the samples were taken.
}

\item{printCSVfile}{
Set to TRUE if an output table has to be produced. The table should be found in (working directory)/Tables/SearchTable.csv. 
}
}
\value{
An output table is stored in the folder (working
directory)/tables/SearchTable.csv if printCSVfile is set to TRUE. More info
at metaboliteTable.
}
\description{
Takes a MAIT object and performs the metabolite search for the significant
features
}
\examples{
\dontrun{
file_name <- system.file("extdata", "peak_table_sig_ann.rds", package = "NIHSlcms")
peak_table <- readRDS(file_name)
metabololite_table <- lcms_identify_metabolites(MAIT.object = peak_table,
                              peakTolerance = 0.005)
}
}
\seealso{
Other metabolite identification functions: 
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_parameter_table}()}

Other dataset_peak_table functions: 
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_tics}()},
\code{\link{lcms_to_mait}()}

Other import/export functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()},
\code{\link{phData}()}
}
\concept{dataset_peak_table functions}
\concept{import/export functions}
\concept{metabolite identification functions}
