% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{lcms_meta_add}
\alias{lcms_meta_add}
\title{Add metadata to MSnExp object}
\usage{
lcms_meta_add(object, metadata, by = "sampleNames")
}
\arguments{
\item{object}{Object, possibly derived from \code{\link{eSet-class}}
    or \code{\linkS4class{AnnotatedDataFrame}}.}

\item{metadata}{A data frame to be merged}

\item{by}{A column present both in \code{metadata} and in \code{Biobase::pData(object)}}
}
\value{
An object from the \link{lcms_dataset_family} with metadata added
}
\description{
Add metadata to MSnExp object
}
\examples{
dataset <- lcms_dataset_load(system.file
                                  ("extdata","dataset.rds",
                                  package = "NIHSlcms"))

metadata <- lcms_meta_read(system.file("extdata",
                                       "metadata.xlsx",
                                       package = "NIHSlcms"))

dataset_metadata <- lcms_meta_add(dataset,
                               metadata,
                               by = "sampleNames")
print(dataset_metadata)
}
\seealso{
Other metadata functions: \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}}, \code{\link{phData}}

Other dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_meta_export}}, \code{\link{phData}}
}
\concept{dataset functions}
\concept{metadata functions}
