% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_read_samples.R
\name{lcms_meta_add}
\alias{lcms_meta_add}
\title{Add metadata to MSnExp object}
\usage{
lcms_meta_add(object, metadata, by = "sampleNames")
}
\arguments{
\item{object}{Object, possibly derived from \code{\link{eSet-class}}
    or \code{\linkS4class{AnnotatedDataFrame}}.}

\item{metadata}{A data frame to be merged}

\item{by}{A column present both in \code{metadata} and in \code{Biobase::pData(object)}}
}
\value{
The object with the added metadata
}
\description{
Add metadata to MSnExp object
}
\examples{
\dontrun{
lcms_dataset <- lcms_dataset_load(system.file("extdata",
                                             "lcms_dataset.rds",
                                              package = "NIHSlcms"))

metadata <- read_excel(system.file("extdata",
                                             "metadata.xlsx",
                                              package = "NIHSlcms"))
lcms_dataset_metadata <- lcms_meta_add(lcms_dataset,metadata)
}

}
