% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_load_save.R
\name{lcms_meta_export}
\alias{lcms_meta_export}
\title{Export Metadata to an Excel file}
\usage{
lcms_meta_export(lcms_dataset, xlsx_file)
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{xlsx_file}{"The .xlsx excel file"}

\item{groups}{A character vector. Use \code{"external"} for the external metadata or
the default for a more generic solution}
}
\value{
The Excel file name
}
\description{
Export Metadata to an Excel file
}
\examples{
\dontrun{
lcms_dataset_metadata <- lcms_dataset_load(system.file("extdata",
                                             "lcms_dataset_metadata.rds",
                                              package = "NIHSlcms"))
xlsx_file <- "exported_metadata.xlsx"
lcms_meta_export(lcms_dataset_metadata, xlsx_file)
}

}
\seealso{
Other metadata functions: \code{\link{lcms_meta_read}}

Other lcms_dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_filterPolarity}},
  \code{\link{lcms_filterRTmin}},
  \code{\link{lcms_filterSampleType}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}}

Other import/export functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_read}}
}
\concept{import/export functions}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
\concept{metadata functions}
