% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_dataset_load_save.R
\name{lcms_meta_export}
\alias{lcms_meta_export}
\title{Export Metadata to an Excel file}
\usage{
lcms_meta_export(lcms_dataset, xlsx_file)
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{xlsx_file}{"The .xlsx excel file"}

\item{groups}{A character vector. Use \code{"external"} for the external metadata or
the default for a more generic solution}
}
\value{
The Excel file name
}
\description{
Export Metadata to an Excel file
}
\seealso{
Other lcms_dataset functions: \code{\link{filterRTmin}},
  \code{\link{filterSampleType}},
  \code{\link{lcms_plot_tics}}, \code{\link{lcms_tics}},
  \code{\link{load_and_save_functions}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_plot_tics}},
  \code{\link{lcms_tics}},
  \code{\link{load_and_save_functions}}

Other import/export functions: \code{\link{load_and_save_functions}}
}
\concept{import/export functions}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
\concept{metadata functions}
