% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{lcms_meta_read}
\alias{lcms_meta_read}
\title{Read metadata}
\usage{
lcms_meta_read(xlsx_file)
}
\arguments{
\item{xlsx_file}{xlsx_file "The .xlsx excel file" with metadata.}
}
\value{
A dataframe with the metadata.
}
\description{
Reads metadata from an Excel file.
}
\examples{
\dontrun{
metadata <- lcms_meta_read(system.file("extdata",
                                             "metadata.xlsx",
                                              package = "NIHSlcms"))
print(metadata[1:6])
}
}
\seealso{
Other metadata functions: 
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{phData}()}

Other import/export functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()},
\code{\link{phData}()}
}
\concept{import/export functions}
\concept{metadata functions}
