% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{lcms_peak_table_boxplots}
\alias{lcms_peak_table_boxplots}
\title{Boxplots for the significant peak table features}
\usage{
lcms_peak_table_boxplots(MAIT.object = NULL, treatment_col)
}
\arguments{
\item{MAIT.object}{MAIT object where it is found an annotated peak table.}

\item{treament_col}{Treatment for the samples.}
}
\value{
BoxPlots are stored in folders associated to their corresponding spectra. No explicit plot is produced by the device.
}
\description{
It performs boxplots for any of significant features found in a peak table. All the
boxplots are stored in a directory /Boxplots/Boxplot_spectra_.
}
\examples{
file_name_1 <-  system.file("extdata","peak_table_sig_ann.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name_1)
file_name_2 <-  system.file("extdata","dataset_pos_rt_rs.rds", package = "NIHSlcms")
dataset <-  base::readRDS(file_name_2)
treatment_col <- scales::hue_pal()(length(unique(dataset$treatment)))
lcms_peak_table_boxplots(peak_table,
                         treatment_col = treatment_col)
}
\seealso{
Other metabolite identification functions: \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_spectral_sig_features}},
  \code{\link{lcms_to_mait}},
  \code{\link{lcms_write_parameter_table}}

Other dataset_peak_table functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_identify_metabolites}},
  \code{\link{lcms_peak_annotation}},
  \code{\link{lcms_peak_table_pca}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_raw_data}},
  \code{\link{lcms_sig_peaks_table}},
  \code{\link{lcms_spectral_sig_features}},
  \code{\link{lcms_tics}}, \code{\link{lcms_to_mait}}

Other statistical functions: \code{\link{lcms_peak_table_pca}}

Other visualization functions: \code{\link{lcms_peak_table_pca}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_chrom}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_retcorgroup_optimization}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lmcs_plot_chrom_peak_image}}
}
\concept{dataset_peak_table functions}
\concept{metabolite identification functions}
\concept{statistical functions}
\concept{visualization functions}
