% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{lcms_peakpicking_optimization}
\alias{lcms_peakpicking_optimization}
\title{Peak picking optimization}
\usage{
lcms_peakpicking_optimization(dataset, peakpickingParameters,
  nSlaves = 4, opt_path, plots = TRUE)
}
\arguments{
\item{peakpickingParameters}{Parameters for peak picking}

\item{nSlaves}{Number of slaves the optimization process should spawn.}

\item{opt_path}{optimization path}

\item{plots}{Defines if plots should be generated (TRUE) or not (FALSE) in a subfolder called "plot_ipo".}

\item{lcms_dataset}{An [lcms_dataset_family] object}
}
\value{
A peak picking list with the best setting
}
\description{
The function optimize parameters considering a set of samples
for the peak picking algorithm using the IPO Package.
This includes Peak Detection (‘Centwave’ and ‘Matched Filter’),
Retention Time Correction (‘obiwarp’) and Peak Correspondence
(‘Density’).
}
\seealso{
Other optimization functions: \code{\link{lcms_corgroup_optimization}},
  \code{\link{lcms_default_corgroup_params}},
  \code{\link{lcms_default_peakpeaking_params}},
  \code{\link{write_opt_params}}
}
\concept{optimization functions}
