% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_parameter_optimization.R
\name{lcms_peakpicking_optimization}
\alias{lcms_peakpicking_optimization}
\title{Peak picking optimization}
\usage{
lcms_peakpicking_optimization(lcms_dataset, peakpickingParameters,
  nSlaves = 1, opt_path, plots = TRUE)
}
\arguments{
\item{lcms_dataset}{A \link{lcms_dataset_family} object}

\item{peakpickingParameters}{Parameters for peak picking}

\item{nSlaves}{Number of slaves the optimization process should spawn.}

\item{opt_path}{Path where optimization samples and optimized parameters are save. If NULL, a temporary folder is created.}

\item{plots}{Defines if plots should be generated (TRUE) or not (FALSE) in a subfolder called "plot_ipo".}
}
\value{
A peak picking list with the best setting
}
\description{
The function optimize parameters considering a set of samples
for the peak picking algorithm using the IPO Package.
This includes Peak Detection (‘Centwave’ and ‘Matched Filter’),
Retention Time Correction (‘obiwarp’) and Peak Correspondence
(‘Density’).
}
\examples{
\dontrun{
file_name <- system.file("extdata", "lcms_dataset_rt_pos_rs.rds", package = "NIHSlcms")
lcms_dataset <- lcms_dataset_load(file_name)
default_peakpicking_params <- lcms_default_peakpicking_params(optimize = TRUE)
resultPeakpicking <- lcms_peakpicking_optimization(lcms_dataset,
                                                   default_peakpicking_params,
                                                   opt_path = NULL)
print(resultPeakpicking)}

}
\seealso{
Other optimization functions: \code{\link{lcms_default_peakpicking_params}},
  \code{\link{lcms_default_retcorgroup_params}},
  \code{\link{lcms_write_opt_params}}
}
\concept{optimization functions}
