% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{lcms_peakpicking_optimization}
\alias{lcms_peakpicking_optimization}
\title{Peak picking optimization}
\usage{
lcms_peakpicking_optimization(dataset, peakpickingParameters,
  nSlaves = 1, opt_path, subdir = "plot_ipo", plots = TRUE)
}
\arguments{
\item{dataset}{A \link{lcms_dataset_family} object}

\item{peakpickingParameters}{Parameters for peak picking}

\item{nSlaves}{Number of slaves the optimization process should spawn.}

\item{opt_path}{Path where optimization samples are saved. Id subdir is}

\item{subdir}{Folder where surface plots are saved. If NULL they are displayed by the graphical device.}

\item{plots}{Defines if plots should be generated (TRUE) or not (FALSE) in a subfolder called "plot_ipo" (default).}
}
\value{
A peak picking list with the best setting
}
\description{
The function optimize parameters considering a set of samples
for the peak picking algorithm using the IPO Package.
This includes Peak Detection (‘Centwave’ and ‘Matched Filter’),
Retention Time Correction (‘obiwarp’) and Peak Correspondence
(‘Density’).
}
\examples{
\dontrun{
opt_path <- system.file("extdata", "ipo_opt", package = "NIHSlcms")
file_name <- system.file("extdata", "dataset_pos_rt_rs.rds", package = "NIHSlcms")
dataset <- lcms_dataset_load(file_name)
default_peakpicking_params <- lcms_default_peakpicking_params(optimize = TRUE)
result_peakpicking <- lcms_peakpicking_optimization(dataset,
                                                   default_peakpicking_params,
                                                   opt_path = opt_path,
                                                   subdir = NULL)
}
}
\seealso{
Other dataset functions: \code{\link{lcms_align_rtime}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_fill_chrom_peaks}},
  \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_find_chrom_peaks_cwp}},
  \code{\link{lcms_group_peaks}},
  \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_rearrange_datafiles_by_class}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lcms_tics}},
  \code{\link{lmcs_plot_chrom_peak_image}},
  \code{\link{phData}}

Other optimization functions: \code{\link{lcms_convert_ipo_to_xcms}},
  \code{\link{lcms_default_peakpicking_params}},
  \code{\link{lcms_default_retcorgroup_params}},
  \code{\link{lcms_read_ipo_to_xcms}},
  \code{\link{lcms_write_opt_params}}

Other visualization functions: \code{\link{lcms_plot_chrom}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_retcorgroup_optimization}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lmcs_plot_chrom_peak_image}}
}
\concept{dataset functions}
\concept{optimization functions}
\concept{visualization functions}
