% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcsm_total_ion_count.R
\name{lcms_plot_tics}
\alias{lcms_plot_tics}
\title{Total Ion Count (TIC) plot}
\usage{
lcms_plot_tics(tics, treatment = treatment, rt = NULL,
  plot_type = "spec")
}
\arguments{
\item{tics}{A Total Ion Count object generated with \code{lcms_tics}}

\item{treatment}{Class groups of the samples}

\item{rt}{Retention time}

\item{plot_type}{The plot class, either boxplot or spectra}
}
\value{
Total Ion Count (TIC) for the polarity samples.
}
\description{
The function performs a plot with the Total Ion Count (TIC).
}
\examples{

lcms_dataset <- lcms_dataset_load(system.file
                                  ("extdata","lcms_dataset_pos.rds",
                                    package = "NIHSlcms"))
tics <- lcms_tics(lcms_dataset)

lcms_plot_tics(tics, treatment = treatment,
               rt = c(4, 14),plot_type = "spec")

lcms_plot_tics(tics, treatment = treatment,
               rt = c(4, 14), plot_type = "boxplot")

}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_filterPolarity}},
  \code{\link{lcms_filterRTmin}},
  \code{\link{lcms_filterSampleType}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}}, \code{\link{lcms_tics}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}}, \code{\link{lcms_tics}}
}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
