% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_peak_chromatogram.R
\name{lcms_plotChrom}
\alias{lcms_plotChrom}
\title{Base peak chromatogram}
\usage{
lcms_plotChrom(chromatogram, treatment_col, rtlim = NULL)
}
\arguments{
\item{chromatogram}{A chromatogram object generated with XCMS}

\item{treatment_col}{Color code by groups. They can be generated using
treatment_col <- scales::hue_pal()(length(unique(lcms_dataset$treatment)))
names(treatment_col) <- unique(lcms_dataset$treatment)}

\item{rtlim}{retention time boundaries (e.g. c(4,14))}
}
\value{
A base peak chromatogram
}
\description{
Base peak chromatograms with retention time axis in minutes.
}
\examples{
rtlim <- c(4, 14)
base_peaks <- xcms::chromatogram(lcms_dataset, aggregationFun = "max")
lcms_plotChrom(base_peaks, treatment_col, rtlim = rtlim)

}
