% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_peak_annotation.R
\name{lcms_raw_data}
\alias{lcms_raw_data}
\title{Raw data extractor from a MAIT object}
\usage{
lcms_raw_data(MAIT.object)
}
\arguments{
\item{MAIT.object}{A MAIT-class object}
}
\value{
A list containing either a xcmsSet or a xsAnnotate object.
}
\description{
Function rawData extracts the raw data used to build the MAIT-class object
}
\examples{
\dontrun{
file_name <-  system.file("extdata", "peak_table_MAIT.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name)

peak_table_ANN <- peak_annotation(MAIT.object = peak_table)
lcms_raw_data(peak_table)
}
}
