% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_converter.R
\name{lcms_raw_to_mzxml}
\alias{lcms_raw_to_mzxml}
\title{Convert RAW samples to mzxml samples}
\usage{
lcms_raw_to_mzxml(samples, options = c("--mzxml", "--select_mono_prec"),
  rawconverter = NULL)
}
\arguments{
\item{samples}{ThermoFisher .raw files}

\item{rawconverter}{The working directory given to \link{install_RawConverter}.}
}
\description{
Convert RAW samples to mzxml samples
}
\section{RawConverter}{


Raw samples from ThermoFisher are not easily parsed. All implementations go
through the proprietary dll from Thermo.

This function uses a tool named RawConverter available at:
http://fields.scripps.edu/rawconv/

RawConverter uses the ThermoFisher tool MSFileReader that takes care of the
decryption.

Installing that tool may not be straightforward on all OS, so we provide some
helpers.

On non-Windows systems:
You will need wine and winetricks installed. Then call \code{install_RawConverter("/your/installation/directory")}
}

