% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_ipo_to_xcms.R
\name{lcms_read_ipo_to_xcms}
\alias{lcms_read_ipo_to_xcms}
\title{Read IPO parameters to XCMS format}
\usage{
lcms_read_ipo_to_xcms(opt_result_path)
}
\arguments{
\item{opt_result_path}{A directory where the parameters file is stored}
}
\value{
A display of the chosen parameters
}
\description{
The function reads and converts from \code{IPO} to \code{XCMS} variable formats: The variable names in IPO and XCMS
presented some mismatches, thus we had to create a function called \emph{lcms_read_ipo_to_xcms}
to achieve compatibility between packages when reading these parameters from a .csv file.
}
\examples{
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
lcms_preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)
print(lcms_preproc_params)

}
