% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{lcms_read_ipo_to_xcms}
\alias{lcms_read_ipo_to_xcms}
\title{Read IPO parameters to XCMS format}
\usage{
lcms_read_ipo_to_xcms(opt_result_path)
}
\arguments{
\item{opt_result_path}{A directory where the parameters file is stored.}
}
\value{
A display of the chosen parameters.
}
\description{
The function reads and converts from \code{IPO} to \code{XCMS} variable formats: The variable names in IPO and XCMS
presented some mismatches, the function \emph{lcms_read_ipo_to_xcms} was created
to achieve compatibility between packages when reading these parameters from a .csv file.
}
\examples{
\dontrun{
opt_result_path <-  system.file("extdata","ipo_opt", package = "NIHSlcms")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)
print(preproc_params)
}
}
\seealso{
Other optimization functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_default_peakpicking_params}()},
\code{\link{lcms_peakpicking_optimization}()},
\code{\link{lcms_retcorgroup_optimization}()},
\code{\link{lcms_write_opt_params}()}

Other import/export functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()},
\code{\link{phData}()}
}
\concept{import/export functions}
\concept{optimization functions}
