% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{lcms_retcorgroup_optimization}
\alias{lcms_retcorgroup_optimization}
\title{Optimization of retention time correction and grouping parameters.}
\usage{
lcms_retcorgroup_optimization(optimizedXcmsSetObject,
  retcorGroupParameters, nSlaves = 1, opt_path, subdir = "plot_ipo",
  plots = TRUE)
}
\arguments{
\item{optimizedXcmsSetObject}{XCMS object conatining the \code{best_settings} parameters.
This object may be created after running \code{lcms_peakpicking_optimization} and extract
best_settings$xset (e.g. optimizedXcmsSetObject <- resultPeakpicking$best_settings$xset)}

\item{retcorGroupParameters}{Parameters for retention time correction and optimization}

\item{nSlaves}{Number of slaves the optimization process should spawn.}

\item{subdir}{Folder where surface plots are save. If NULL they are displayed by the graphical device.}

\item{plots}{Defines if plots should be generated (TRUE) or not (FALSE) in a
subfolder called "plot_ipo"(default).}
}
\value{
a list with the optimization of parameters for retention time and grouping.
}
\description{
The function optimizes parameters considering a set of samples
for the retention time correction and grouping using the IPO Package.
}
\examples{
\dontrun{
file_name_pp <- system.file("extdata", "result_peakpicking.rds", package = "NIHSlcms")
optimizedXcmsSetObject <-base::readRDS(file_name_pp)$best_settings$xset
file_name_rcg <- system.file("extdata", "default_retcorgroup_params.rds", package = "NIHSlcms")
default_retcorgroup_params <- base::readRDS(file_name_rcg)
opt_path <-  system.file("extdata", package = "NIHSlcms")

result_retcorgroup <- lcms_retcorgroup_optimization(optimizedXcmsSetObject,
                                                   default_retcorgroup_params,
                                                   opt_path = opt_path,
                                                   subdir = NULL)
}
}
\seealso{
Other optimization functions.: \code{\link{lcms_default_retcorgroup_params}}

Other visualization functions: \code{\link{lcms_peak_table_boxplots}},
  \code{\link{lcms_peak_table_pca}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_chrom}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lmcs_plot_chrom_peak_image}}
}
\concept{optimization functions.}
\concept{visualization functions}
