% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_retention_time_alignment_plot}
\alias{lcms_retention_time_alignment_plot}
\title{Retention Time Correction Plot}
\usage{
lcms_retention_time_alignment_plot(data)
}
\arguments{
\item{data}{An aligned lcms_dataset}
}
\value{
The plot for the retention time correction
}
\description{
It plots the retention time correction vs the original retention time for each of the samples
coloured by sample class.
}
\examples{
file_name <-  system.file("extdata", "peakdet_align.rds", package = "NIHSlcms")
data <- base::readRDS(file_name)
rta_plot <- lcms_retention_time_alignment_plot(data)
print(rta_plot)
}
