% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{lcms_spectral_sig_features}
\alias{lcms_spectral_sig_features}
\title{Extract significant features from a MAIT object}
\usage{
lcms_spectral_sig_features(
  MAIT.object = NULL,
  pvalue = 0.05,
  p.adj = "none",
  printCSVfile = FALSE,
  scale = FALSE,
  parametric = TRUE,
  var.equal = FALSE,
  test.fun = NULL,
  jitter = FALSE,
  jitter.factor = 1,
  jitter.amount = 0,
  namefun = NULL
)
}
\value{
A MAIT-class object containing the significant features of the scores
slot of MAIT-class object used as an input.
}
\description{
Function lcms_spectral_sig_features takes a MAIT-class object and obtains which of
the variables are significant given a p-value threshold. The parameters of
the significant features can ve printed to an output table (TRUE by default).
Depending on the number of classes in the data, the function chooses between
using ANOVA test or T-Student test.
}
\examples{
\dontrun{
file_name <-  system.file("extdata", "peak_table_ann.rds", package = "AlpsLCMS")
peak_table <- base::readRDS(file_name)
peak_table_sig_ann <- lcms_spectral_sig_features(MAIT.object = peak_table,
                                          pvalue=0.05,
                                          p.adj="none",
                                          scale=FALSE)
print(peak_table_sig_ann)
}
}
\seealso{
Other metabolite identification functions: 
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_parameter_table}()}

Other dataset_peak_table functions: 
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_tics}()},
\code{\link{lcms_to_mait}()}

Other import/export functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()},
\code{\link{phData}()}
}
\concept{dataset_peak_table functions}
\concept{import/export functions}
\concept{metabolite identification functions}
