% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lcms_tics}
\alias{lcms_tics}
\title{Total Ion Count (TIC)}
\usage{
lcms_tics(lcms_dataset, treatment = treatment)
}
\arguments{
\item{lcms_dataset}{An \link{lcms_dataset_family} object}

\item{treatment}{Class groups of the samples}
}
\value{
Total Ion Count (TIC) for the polarity samples.
}
\description{
The function performs the Total Ion Count (TIC) for the polarity samples.
Function \code{lcms_tics} stores summarizes TIC information.
NOTE: \code{lcms_tics} assumes that data is already filtered by polarity.
}
\examples{
dataset <- lcms_dataset_load(system.file
                                  ("extdata","dataset_pos.rds",
                                    package = "NIHSlcms"))
tics <- lcms_tics(dataset)


print(tics)
}
\seealso{
Other lcms_dataset functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}

Other lcms_dataset_peak_table functions: \code{\link{lcms_dataset_load}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_meta_read}},
  \code{\link{lcms_plot_tics}}
}
\concept{lcms_dataset functions}
\concept{lcms_dataset_peak_table functions}
