% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_MAIT.R
\name{lcms_to_MAIT}
\alias{lcms_to_MAIT}
\title{to MAIT}
\usage{
lcms_to_MAIT(dataDir = NULL, projectDir = NULL, project = NULL,
  preproc_params = NULL, peakTable = NULL)
}
\arguments{
\item{dataDir}{directory with LC-MS datafiles}

\item{projectDir}{path to the the project directory}

\item{project}{name of the}

\item{preproc_params}{params from XCMS}

\item{peakTable}{XCMS procesed LC-MS data}
}
\value{
A MAIT object
}
\description{
Function to create a MAIT object using data from xcms
}
\examples{

file_name <-  system.file("extdata", "peak_table_imputed.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name)
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
lcms_preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

dataDir <-  system.file("extdata", "rearrange_mait", package = "NIHSlcms")
projectDir <-  system.file("extdata", package = "NIHSlcms")
project <- "project"

peak_table_MAIT <- lcms_to_MAIT(dataDir = dataDir,
                                projectDir = projectDir,
                                project = project,
                                preproc_params = lcms_preproc_params,
                                peakTable = peak_table)
print(peak_table_MAIT)
}
