% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{lcms_to_mait}
\alias{lcms_to_mait}
\title{Convert xcms to MAIT}
\usage{
lcms_to_mait(
  data_dir = NULL,
  project_dir = NULL,
  project = NULL,
  preproc_params = NULL,
  peak_table = NULL
)
}
\arguments{
\item{data_dir}{directory with LC-MS datafiles.}

\item{project_dir}{path to the the project directory.}

\item{project}{name of the.}

\item{preproc_params}{params from XCMS.}

\item{peak_table}{XCMS procesed LC-MS data.}
}
\value{
A MAIT object.
}
\description{
Function to create a MAIT object using data from xcms.
}
\details{
\code{MAIT} Package allows to annotate the peaks of the peak table provided by \code{XCMS}.
*Note: The dataset must be converted from an object of the XCMS Package
to an object of the MAIT Package.
}
\examples{
\dontrun{
file_name <-  system.file("extdata", "peak_table_imputed.rds", package = "AlpsLCMS")
peak_table <- base::readRDS(file_name)
opt_result_path <-  system.file("extdata", package = "AlpsLCMS")
preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

data_dir <-  system.file("extdata", "rearrange_mait", package = "AlpsLCMS")
project_dir <-  system.file("extdata", package = "AlpsLCMS")
project <- "project"

peak_table_mait <- lcms_to_mait(data_dir = data_dir,
                                project_dir = project_dir,
                                project = project,
                                preproc_params = preproc_params,
                                peak_table = peak_table)
print(peak_table_mait)
}
}
\seealso{
Other metabolite identification functions: 
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_write_parameter_table}()}

Other dataset_peak_table functions: 
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_peak_table_boxplots}()},
\code{\link{lcms_peak_table_pca}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_tics}()}

Other import/export functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()},
\code{\link{phData}()}
}
\concept{dataset_peak_table functions}
\concept{import/export functions}
\concept{metabolite identification functions}
