% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_MAIT.R
\name{lcms_writeParameterTable}
\alias{lcms_writeParameterTable}
\title{Write Parameters Table for MAIT functionalities}
\usage{
lcms_writeParameterTable(listParameters, folder)
}
\arguments{
\item{listParameters}{Parameters}

\item{folder}{A directory to store a .csv file where the  parameter table is stored.}
}
\value{
A template with MAIT parameters.
}
\description{
Write Parameters Table for MAIT functionalities
}
\examples{
\dontrun{
dataDir <-  system.file("extdata", "rearrange_mait", package = "NIHSlcms")
MAITparams_path <- system.file("extdata", "Results_project", package = "NIHSlcms")
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
lcms_preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)
parameters <- c(dataDir = dataDir, lcms_preproc_params)

quiet <- function(x) {
             base::sink(base::tempfile())
             base::on.exit(base::sink())
             base::invisible(base::force(x))}

quiet(getClassDef("MAIT","MAIT"))
MAIT.object = signature(MAIT.Object = "MAIT")
MAIT.object <- methods::new("MAIT")
MAIT.object@RawData@parameters@sampleProcessing <- parameters
lcms_writeParameterTable(MAIT::parameters(MAIT.object), folder = MAITparams_path)
}

}
