% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{lmcs_plot_chrom_peak_image}
\alias{lmcs_plot_chrom_peak_image}
\title{Image of Chromatographic Peaks by sample}
\usage{
lmcs_plot_chrom_peak_image(x, binSize = 30, xlim = NULL, log = FALSE,
  xlab = "retention time", yaxt = par("yaxt"),
  main = "Chromatographic peak counts", ...)
}
\arguments{
\item{x}{A lcms_dataset.}

\item{binSize}{Bin size for the the reteincion time axis.}

\item{xlim}{Retention time axis range. If Null, it uses the complete retention time range.}

\item{log}{If TRUE, it performs a logarithmic transformation on the data.}

\item{xlab}{Label for the retention time axis.}

\item{yaxt}{Options for the y axis. Use the default value.}

\item{...}{Additional parameters for image function.}
}
\value{
An image with the detected chromatographic peak, for each sample.
}
\description{
It plots the an image of the chromatographic peaks for each sample. This function is useful if
you are interested in knowing the effect of the retention time correction on the chromatographic axis.
}
\examples{
file_name <-  system.file("extdata", "peakdet_align.rds", package = "NIHSlcms")
dataset <- base::readRDS(file_name)
chr_peak_image <- lmcs_plot_chrom_peak_image(dataset, binSize = 5, xlim = NULL, log = FALSE,
                                               xlab = "retention time (min)", yaxt = par("yaxt"))
                                               title(main ="Detected Peaks (Aligned)")
}
\seealso{
Other dataset functions: \code{\link{lcms_align_rtime}},
  \code{\link{lcms_dataset_load}},
  \code{\link{lcms_dataset_save}},
  \code{\link{lcms_fill_chrom_peaks}},
  \code{\link{lcms_filter_mz}},
  \code{\link{lcms_filter_polarity}},
  \code{\link{lcms_filter_rt_min}},
  \code{\link{lcms_filter_sample_type}},
  \code{\link{lcms_find_chrom_peaks_cwp}},
  \code{\link{lcms_group_peaks}},
  \code{\link{lcms_meta_add}},
  \code{\link{lcms_meta_export}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_rearrange_datafiles_by_class}},
  \code{\link{lcms_retention_time_alignment_plot}},
  \code{\link{lcms_tics}}, \code{\link{phData}}

Other chromatogram functions: \code{\link{lcms_plot_chrom}},
  \code{\link{lcms_plot_tics}}

Other visualization functions: \code{\link{lcms_peak_table_boxplots}},
  \code{\link{lcms_peak_table_pca}},
  \code{\link{lcms_peakpicking_optimization}},
  \code{\link{lcms_plot_chrom}},
  \code{\link{lcms_plot_tics}},
  \code{\link{lcms_retcorgroup_optimization}},
  \code{\link{lcms_retention_time_alignment_plot}}
}
\concept{chromatogram functions}
\concept{dataset functions}
\concept{visualization functions}
