% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{PeakGroupsPar}
\alias{PeakGroupsPar}
\title{Peak density parameters}
\usage{
PeakGroupsPar(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[xcms:adjustRtime-peakGroups]{xcms::PeakGroupsParam}}
  \describe{
    \item{\code{minFraction}}{\code{numeric(1)} between 0 and 1 defining the minimum
required fraction of samples in which peaks for the peak group were
identified. Peak groups passing this criteria will aligned across
samples and retention times of individual spectra will be adjusted
based on this alignment. For \code{minFraction = 1} the peak group
has to contain peaks in all samples of the experiment. Note that if
\code{subset} is provided, the specified fraction is relative to the
defined subset of samples and not to the total number of samples within
the experiment (i.e. a peak has to be present in the specified
proportion of subset samples).}
    \item{\code{extraPeaks}}{\code{numeric(1)} defining the maximal number of
additional peaks for all samples to be assigned to a peak group (i.e.
feature) for retention time correction. For a data set with 6 samples,
\code{extraPeaks = 1} uses all peak groups with a total peak count
\code{<= 6 + 1}. The total peak count is the total number of peaks being
assigned to a peak group and considers also multiple peaks within a
sample being assigned to the group.}
    \item{\code{smooth}}{character defining the function to be used, to interpolate
corrected retention times for all peak groups. Either \code{"loess"} or
\code{"linear"}.}
    \item{\code{span}}{\code{numeric(1)} defining the degree of smoothing (if
\code{smooth = "loess"}). This parameter is passed to the internal call
to \code{\link{loess}}.}
    \item{\code{family}}{character defining the method to be used for loess smoothing.
Allowed values are \code{"gaussian"} and \code{"symmetric"}.See
\code{\link{loess}} for more information.}
    \item{\code{peakGroupsMatrix}}{optional \code{matrix} of (raw) retention times for
the peak groups on which the alignment should be performed. Each column
represents a sample, each row a feature/peak group. Such a matrix is
for example returned by the \code{\link[xcms]{adjustRtimePeakGroups}} method.}
    \item{\code{subset}}{\code{integer} with the indices of samples within the
experiment on which the alignment models should be estimated. Samples
not part of the subset are adjusted based on the closest subset sample.
See description above for more details.}
    \item{\code{subsetAdjust}}{\code{character} specifying the method with which
non-subset samples should be adjusted. Supported options are
\code{"previous"} and \code{"average"} (default). See description above
for more information.}
  }}
}
\value{
PeakGroupsParam S4 object
}
\description{
The method creates a template with the parameters for \strong{alignment}.
Extracted from the function \code{xcms::PeakGroupsParam}.
}
\section{Slots}{


\describe{
\item{\code{.__classVersion__,minFraction,extraPeaks,smooth,span,family,peakGroupsMatrix,subset,subsetAdjust}}{See corresponding parameter above. \code{.__classVersion__} stores
the version from the class. Slots values should exclusively be accessed
\emph{via} the corresponding getter and setter methods listed above.}
}
}

\examples{
##############################
## Chromatographic peak detection and grouping.
##
## Below we perform first a peak detection (using the matchedFilter
## method) on some of the test files from the faahKO package followed by
## a peak grouping.
library(faahKO)
library(xcms)
fls <- dir(system.file("cdf/KO", package = "faahKO"), recursive = TRUE,
           full.names = TRUE)

## Reading 2 of the KO samples
raw_data <- readMSData(fls[1:2], mode = "onDisk")

## Perform the peak detection using the matchedFilter method.
mfp <- MatchedFilterParam(snthresh = 20, binSize = 1)
res <- findChromPeaks(raw_data, param = mfp)

head(chromPeaks(res))
## The number of peaks identified per sample:
table(chromPeaks(res)[, "sample"])

## Performing the peak grouping using the "peak density" method.
p <- PeakDensityParam(sampleGroups = c(1, 1))
res <- groupChromPeaks(res, param = p)

## Perform the retention time adjustment using peak groups found in both
## files.
fgp <- PeakGroupsParam(minFraction = 1)

## Before running the alignment we can evaluate which features (peak groups)
## would be used based on the specified parameters.
pkGrps <- adjustRtimePeakGroups(res, param = fgp)

## We can also plot these to evaluate if the peak groups span a large portion
## of the retention time range.
plot(x = pkGrps[, 1], y = rep(1, nrow(pkGrps)), xlim = range(rtime(res)),
    ylim = c(1, 2), xlab = "rt", ylab = "", yaxt = "n")
points(x = pkGrps[, 2], y = rep(2, nrow(pkGrps)))
segments(x0 = pkGrps[, 1], x1 = pkGrps[, 2],
    y0 = rep(1, nrow(pkGrps)), y1 = rep(2, nrow(pkGrps)))
grid()
axis(side = 2, at = c(1, 2), labels = colnames(pkGrps))

## Next we perform the alignment.
res <- adjustRtime(res, param = fgp)

## Any grouping information was dropped
hasFeatures(res)

## Plot the raw against the adjusted retention times.
plot(rtime(raw_data), rtime(res), pch = 16, cex = 0.25, col = fromFile(res))

## Adjusterd retention times can be accessed using
## rtime(object, adjusted = TRUE) and adjustedRtime
all.equal(rtime(res), adjustedRtime(res))

## To get the raw, unadjusted retention times:
all.equal(rtime(res, adjusted = FALSE), rtime(raw_data))

## To extract the retention times grouped by sample/file:
rts <- rtime(res, bySample = TRUE)
}
\references{
Colin A. Smith, Elizabeth J. Want, Grace O'Maille, Ruben Abagyan and
Gary Siuzdak. "XCMS: Processing Mass Spectrometry Data for Metabolite
Profiling Using Nonlinear Peak Alignment, Matching, and Identification"
\emph{Anal. Chem.} 2006, 78:779-787.
}
\seealso{
The \code{\link[xcms]{do_adjustRtime_peakGroups}} core
    API function and \code{\link[xcms]{retcor.peakgroups}} for the old user
    interface.
    \code{\link[xcms]{plotAdjustedRtime}} for visualization of alignment results.

\code{\link[xcms]{XCMSnExp}} for the object containing the results of
    the alignment.

Other retention time correction methods: \code{\link[xcms]{adjustRtime-obiwarp}},
  \code{\link[xcms]{adjustRtime}}
}
\author{
Colin Smith, Johannes Rainer
}
