% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_management.R
\name{phData}
\alias{phData}
\title{Generic function to access the phenotypic data (Biobase)}
\usage{
phData(object)
}
\arguments{
\item{object}{Object, possibly derived from \code{\link[Biobase]{eSet-class}}
    or \code{\linkS4class{AnnotatedDataFrame}}.}
}
\value{
phenoData returns an object containing information on
both variable values and variable meta-data. varLabels returns
a character vector of measured variables. pData returns a data
frame with samples as rows, variables as columns. varMetadata
returns a data frame with variable names as rows, description
tags (e.g., unit of measurement) as columns.
}
\description{
It uses \link[Biobase:phenoData]{Biobase::pData} to access phenotypic data such as
metadata, covariates, etc...
}
\examples{
\dontrun{
dataset <- lcms_dataset_load(system.file("extdata",
                                         "dataset.rds",
                                          package = "AlpsLCMS"))
xlsx_file <- system.file("extdata", "exported_metadata.xlsx", package = "AlpsLCMS")
metadata<- readxl::read_excel(xlsx_file)
dataset_metadata <- lcms_meta_add(dataset, metadata, by = "sampleNames")

head(phData(dataset_metadata))[1:4]
}
}
\seealso{
Other metadata functions: 
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()}

Other dataset functions: 
\code{\link{IPO_group_peaks}()},
\code{\link{find_peaks_cwp}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_fill_chrom_peaks}()},
\code{\link{lcms_filter_mz}()},
\code{\link{lcms_filter_polarity}()},
\code{\link{lcms_filter_rt_min}()},
\code{\link{lcms_filter_sample_type}()},
\code{\link{lcms_meta_add}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_plot_chrom_peak_image}()},
\code{\link{lcms_plot_tics}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_retention_time_alignment_plot}()},
\code{\link{lcms_tics}()},
\code{\link{obiAdjust_Rtime}()},
\code{\link{peakpicking_optimization}()}

Other import/export functions: 
\code{\link{lcms_convert_ipo_to_xcms}()},
\code{\link{lcms_dataset_load}()},
\code{\link{lcms_dataset_save}()},
\code{\link{lcms_identify_metabolites}()},
\code{\link{lcms_meta_export}()},
\code{\link{lcms_meta_read}()},
\code{\link{lcms_peak_annotation}()},
\code{\link{lcms_raw_data}()},
\code{\link{lcms_read_ipo_to_xcms}()},
\code{\link{lcms_read_samples}()},
\code{\link{lcms_rearrange_datafiles_by_class}()},
\code{\link{lcms_sig_peaks_table}()},
\code{\link{lcms_spectral_sig_features}()},
\code{\link{lcms_to_mait}()},
\code{\link{lcms_write_opt_params}()},
\code{\link{lcms_write_parameter_table}()}
}
\concept{dataset functions}
\concept{import/export functions}
\concept{metadata functions}
