% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phData.R
\name{phData}
\alias{phData}
\title{Generic function to access the phenotypic data (Biobase)}
\usage{
phData(object)
}
\arguments{
\item{object}{a \code{lcms_dataset}}
}
\value{
phenoData returns an object containing information on
both variable values and variable meta-data. varLabels returns
a character vector of measured variables. pData returns a data
frame with samples as rows, variables as columns. varMetadata
returns a data frame with variable names as rows, description
tags (e.g., unit of measurement) as columns.
}
\description{
It uses \link[Biobase:pData]{Biobase::pData} to access phenotypic data such as
metadata, covariates, etc...
}
\examples{
\dontrun{
metadata<- readxl::read_excel("D:/myFile.xlsx")
dataset <- lcms_meta_add(dataset, metadata, by = "sampleNames")
phData(dataset)
}
}
