% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{write_opt_params}
\alias{write_opt_params}
\title{Displaying and Storing optimized settings}
\usage{
write_opt_params(results_pp, results_rtcg, opt_result_path, csv = TRUE,
  console = TRUE)
}
\arguments{
\item{opt_result_path}{A directory to save the parameters file}

\item{console}{}
}
\description{
The function allows visulizing the parameter optimization results by `IPO` Package
allows in the RStudio console. Also you can save this results in plain text files
(i.e. a .CVS file).
}
\examples{
opt_result_path <- "C:/my_directory"
write_opt_params(resultPeakpicking, resultRetcorGroup, opt_result_path)
}
\seealso{
Other optimization functions: \code{\link{lcms_corgroup_optimization}},
  \code{\link{lcms_default_corgroup_params}},
  \code{\link{lcms_default_peakpeaking_params}},
  \code{\link{lcms_peakpicking_optimization}}
}
\concept{optimization functions}
